% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reanno.R
\name{check_reanno}
\alias{check_reanno}
\title{Checks annotation from reannotation mapping}
\usage{
check_reanno(anno, reanno, threads = 1)
}
\arguments{
\item{anno}{PAC-list object or annotation data.frame with sequences as row
names.}

\item{reanno}{reannotation list generated by \code{import_reanno}.}

\item{threads}{Integer stating the number of parallell jobs.}
}
\value{
logical vector stating if an anno sequences is present (TRUE)  or
  absent (FALSE) in reanno.
}
\description{
\code{check_reanno} check if names are available in reanno import.
}
\details{
Given the path to bowtie mapping output, \code{\link{import_reanno}} will attempt to
read these files into R and generate a list of unordered data frames where
each row summarizes all annotations for a given sequence. \code{check_reanno}
will check if sequences are listed in any of the imported reannotation
data frames. \code{\link{add_reanno}} will add the new annotations to a PAC-list
object.
}
\examples{
#reanno <- import_reanno(bowtie_path=<your_path_to_bowtie_output_directory>, threads=8)
#reanno <- import_reanno(bowtie_path="/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/reanno", threads=8)
#
#bowtie_path <- "/data/Data_analysis/Projects/Drosophila/Specific_projects/Test_temp/Processed_Sports_09-10-19/R_files"
}
\seealso{
\url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
  Bowtie.

\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC reannotation: 
\code{\link{add_reanno}()},
\code{\link{import_reanno}()},
\code{\link{make_reanno}()},
\code{\link{map_reanno}()},
\code{\link{simplify_reanno}()}
}
\concept{PAC reannotation}
