% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{colnames,PAC-method}
\alias{colnames,PAC-method}
\title{sample names}
\usage{
\S4method{colnames}{PAC}(x)
}
\arguments{
\item{x}{A S4 object of class PAC.}
}
\value{
The sample names held by the Counts table (column names) and the
  Pheno table (row names) of a PAC-object.
}
\description{
sample names
}
\examples{

load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))

# extra fuctionality with s4 PAC-object:
names(pac)
length(pac)
nrow(pac)
ncol(pac)
rownames(pac)
colnames(pac)
pheno(pac) 
head(anno(pac))
head(counts(pac))
head(norm(pac)$cpm)

}
