% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sports.R
\name{import_sports}
\alias{import_sports}
\title{Imports Sports Output Files Into R}
\usage{
import_sports(path, threads = 1)
}
\arguments{
\item{path}{A character vector with the path to Sport output directory.}

\item{threads}{An integer vector stating the number of parallell jobs.}
}
\value{
The function will search the Sports output directory (path) for all files
  ending with _output.txt and import them into a list of data.frames.

  Using threads > 1, files will be imported in parallell, one file per
  dedicated thread. Use \code{parallel::detectcores()} to see available
  threads on the computer. Do not exceed the number of available threads!
}
\description{
\code{import_sports} uses parallel processing to read Sports result files and
store them in a list of dataframes.
}
\details{
Given the path to Sports output directory, this function will import all
result output.txt files of a given Sports project and store them in a list of
data.frames. This is done while perserving the basenames of the original
fastq file (input file for Sports). Files will be imported
}
\examples{
  sports_lst <- import_sports(<your_path_to_sports_output_directory>, threads=<number_of_jobs>)
  sports_lst <- import_sports(path=path, threads=12)

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports. \url{https://github.com/Danis102} for
  updates on the current package.

Other PAC generation: \code{\link{make_PAC}},
  \code{\link{make_anno}}, \code{\link{make_counts}},
  \code{\link{make_pheno}}, \code{\link{progress_report}}
}
\concept{PAC generation}
