% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{length,PAC-method}
\alias{length,PAC-method}
\alias{length,reanno-method}
\title{number of objects in PAC}
\usage{
\S4method{length}{PAC}(x)

\S4method{length}{reanno}(x)
}
\description{
number of objects in PAC

length of reanno object
}
