% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_map.R
\name{PAC_map}
\alias{PAC_map}
\title{Aligning sequences against references to produce annotation information}
\usage{
PAC_map(
  PAC,
  output = NULL,
  input_genome,
  input_biotype = NULL,
  hierarchy = NULL,
  bio_search = NULL,
  mismatches = 3,
  filter_hit = TRUE,
  save_name = FALSE,
  threads = 1,
  override = TRUE,
  ...
)
}
\arguments{
\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names.}

\item{output}{Character indicating the path to the destination folder
for the temporary files generated by the function. If no output folder is 
defined, files will be stored in a temporary folder and may be removed on 
system restart.}

\item{input_genome}{List of file paths (character) indicating the full paths to
each fasta reference file used for genome mapping. Note: All reference fasta 
files must have corresponding bowtie indexes. If none exist, you may use
\code{Rbowtie::bowtie_build} to produce them.}

\item{input_biotype}{List of file paths (character) indicating the full paths to
each fasta reference file used for biotype mapping. The names in this list 
determines which names will be used in the final annotation. Note: All 
reference fasta files must have bowtie indexes using \code{Rbowtie::bowtie_build}.}

\item{hierarchy}{list of which hierarchy to use for biotype assignment with
\code{\link{simplify_reanno}}, where more details can be found on its construction. 
If empty, the function will provide a generalized hierarchy based off the biotypes 
found in the data following this order: rRNA -> miRNA -> tRNA -> piRNA -> all other.}

\item{bio_search}{List of character vectors indicating search terms targeting
feature names in the original reference fasta files. The function will look for
the bio_search terms with \code{\link{grepl}} and annotate accordingly. For
more details on how this works, please see \code{\link{add_reanno}}.
If bio_search is NULL, this function will try to estimate which biotypes that are present
in references used by looking for characters combined with "RNA", e.g "miRNA". 
This assumes the references used have named sequences in such a fashion, and may 
work subpar with some references due to that.}

\item{mismatches}{Integer indicating the maximum number of mismatches allowed
in the alignments. The function currently supports no more than 3
mismatches (botwie max). Default=3.}

\item{filter_hit}{If filter_hit is TRUE, sequences without a hit to the reference
genome(s) defined in input_genome up until the mismatches defined will be removed.
Default is TRUE.}

\item{save_name}{logical statement whether to report back full length names of
biotype annotations, useful for further annotations such as feature-based annotations. 
The default is set to FALSE, as this may create Anno objects of wonky dimensions 
with certain reference .fasta files.}

\item{threads}{Integer indicating the number of parallel processes to be
used. Default=1.}

\item{override}{logical whether to write over files in output folder (TRUE),
or whether to ask user for console input to confirm overwriting files. Default
for this function is TRUE, so be cautious if defining output folder!}

\item{...}{Arguments to be passed on to \code{\link{map_reanno}}, \code{\link{make_reanno}},
\code{\link{add_reanno}}, and \code{\link{simplify_reanno}}.}
}
\value{
Will return a PAC object with an expanded Anno.
}
\description{
Aligns sequences in a PAC list object against specified references, returning
a data frame or PAC object containing annotation information
}
\details{
Given a PAC object, the destination of reference files, and information of
what type of information to save, this function will perform mapping against references
and retain information of mapping. This function is based of three main functions;
 \code{map_reanno}, \code{make_reanno} and \code{add_reanno}, whom may be ran 
 separately for more detailed control of the workflow. 
 Using save_name=TRUE in this function will retain the name of the references where sequences are found,
 which may cause large objects.
}
\examples{

#Example for general mapping, for more advanced uses please see
#?map_reanno, ?simply_reanno an ?add_reanno

#Ensure we have the mapping references and their bowtie indexes
## genome (here mycoplasma is used due to its small genome size):

mycoplasma_file <- system.file("extdata/mycoplasma_genome",
                               "mycoplasma.fa",
                               package = "seqpac", mustWork = TRUE)
mycoplasma_dir<- gsub("mycoplasma.fa", "", mycoplasma_file)

if(!sum(stringr::str_count(list.files(mycoplasma_dir), ".ebwt")) ==6){
  Rbowtie::bowtie_build(mycoplasma_file,
                        outdir=mycoplasma_dir,
                        prefix="mycoplasma", force=TRUE)
}

input_genome <- list(genome=mycoplasma_file)

## tRNA:
trna_file <- system.file("extdata/trna", "tRNA.fa",
                         package = "seqpac", mustWork = TRUE)
trna_dir <- dirname(trna_file)

if(!sum(stringr::str_count(list.files(trna_dir), ".ebwt")) ==6){
  Rbowtie::bowtie_build(trna_file,
                        outdir=trna_dir,
                        prefix="tRNA", force=TRUE)
}
## rRNA:
rrna_file <- system.file("extdata/rrna", "rRNA.fa",
                         package = "seqpac", mustWork = TRUE)
rrna_dir <- dirname(rrna_file)

if(!sum(stringr::str_count(list.files(rrna_dir), ".ebwt")) ==6){
  Rbowtie::bowtie_build(rrna_file,
                        outdir=rrna_dir,
                        prefix="rRNA", force=TRUE)
}

input_biotype <- list(trna= trna_file, rrna= rrna_file)

##  load a PAC-object and remove previous mapping from anno:
load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata",
                 package = "seqpac", mustWork = TRUE))
anno(pac) <- anno(pac)[,1, drop = FALSE]

pac_anno <- PAC_map(PAC=pac, input_genome=input_genome,
                     input_biotype=input_biotype, filter_hit = FALSE) 

}
\seealso{
\url{http://bowtie-bio.sourceforge.net/index.shtml} for information
  about Bowtie and for Rbowtie:
  \url{https://www.bioconductor.org/packages/release/bioc/html/Rbowtie.html}.
  \url{https://github.com/OestLab/seqpac} for updates on the current package.

Other PAC reannotation:
\code{\link[=PAC_annotate]{PAC_annotate()}},
\code{\link[=add_reanno]{add_reanno()}},
\code{\link[=as.reanno]{as.reanno()}},
\code{\link[=import_reanno]{import_reanno()}},
\code{\link[=make_conv]{make_conv()}},
\code{\link[=make_reanno]{make_reanno()}},
\code{\link[=map_reanno]{map_reanno()}},
\code{\link[=simplify_reanno]{simplify_reanno()}}
}
\concept{PAC reannotation}
