% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_rpm.R
\name{PAC_rpm}
\alias{PAC_rpm}
\title{Generates RPM values from a PAC object}
\usage{
PAC_rpm(PAC)
}
\arguments{
\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names and a Count table with raw counts.}
}
\value{
dataframe
}
\description{
\code{generate_rpm} Generates RPM values from a PAC object
}
\details{
Using the counts in a PAC object to generate RPM values in a dataframe with
the same rownames as in the original PAC object
}
\examples{
  df  <- PAC_rpm(PAC) 
}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: \code{\link{PAC_covplot}},
  \code{\link{PAC_deseq}}, \code{\link{PAC_filter}},
  \code{\link{PAC_jitter}}, \code{\link{PAC_mapper}},
  \code{\link{PAC_mismap}}, \code{\link{PAC_nbias}},
  \code{\link{PAC_pca}}, \code{\link{PAC_pie}},
  \code{\link{PAC_saturation}}, \code{\link{PAC_sizedist}},
  \code{\link{PAC_stackbar}}, \code{\link{PAC_summary}},
  \code{\link{PAC_trna}}
}
\concept{PAC analysis}
