% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_reanno.R
\name{add_reanno}
\alias{add_reanno}
\title{Imports annotation from reannotation mapping}
\usage{
add_reanno(reanno_path, PAC = NULL, mis_fasta_check = FALSE)
}
\arguments{
\item{reanno_path}{Path to a directory where reannotation .Rdata files can be
found.}

\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names.}

\item{mis_fasta_check}{Logical TRUE/FALSE if checking against anno_misX.fa
should be done.}
}
\value{
Data.frame with additional information from reannotation using
  bowtie. If \emph{PAC} is specified, the function will return the specifed
  PAC object with an added reannotation column in the Anno data.frame.
  If \emph{mis_fasta_check} is specified the function will look for a
  \emph{anno_misX.fa} (X = the file with the highest number) previoulsy
  generated by the reannotation workflow. This file is used to double check
  so that no sequences are missing before and after reannotation.
}
\description{
\code{import_reanno} adds reannotation to PAC.
}
\details{
Given the path to reannotation files (reanno_mis0/1/2/3/4/5.Rdata) generated
by functions in the PAC reannotation family, this function will summarize the
reannotation files into one output. Optionally a PAC object can be provided,
in which case the function will attempt to merge the new reannotation to the
existing Anno data.frame of the PAC object.
}
\examples{
reanno <- add_reanno(reanno_path, PAC=PAC, mis_fasta_check=TRUE)    # Complete use

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports.
  \url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
  Bowtie \url{https://github.com/Danis102} for updates on the current
  package.

Other PAC reannotation: \code{\link{check_reanno}},
  \code{\link{import_reanno}}
}
\concept{PAC reannotation}
