% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_reanno.R
\name{add_reanno}
\alias{add_reanno}
\title{Generates biotype annotation from reannotation object}
\usage{
add_reanno(reanno, mismatches = 0, type = "genome", bio_search, max_hits = 10)
}
\arguments{
\item{reanno}{A reannotation list object generated by \emph{make_reanno}, with
an Overview dataframe and a Full_anno list.}

\item{mismatches}{Integer indicating the number of mismatches that should be
reported. Can never have a higher number than was originally used in the
reannotation workflow. While default=0, reporting the maximum number is
recommended. The mismatch information is only added to the report, not
exchanging annotation hits with less mismatches.}

\item{bio_search}{List of character vectors indicating search terms targeting
feature names in the original reference fasta files.}
}
\value{
Dataframe with mismatch and biotype information (search term hits)
  that can directly be added to an Anno dataframe of a PAC object, given that
  the same PAC object was used in the reannotation workflow.
}
\description{
\code{add_reanno} adds biotypes from reannotation to PAC.
}
\details{
Given a reannotation object generated by \emph{make_reanno} and biotype search
list (see below), information biotype and mismatch information is extracted
from the reannotation object and compiled into a dataframe with the same
sequence order as in the original PAC master file.
}
\examples{
library(seqpac)
load(system.file("extdata", "drosophila_sRNA_pac.Rdata", package = "seqpac", mustWork = TRUE))

outpath_genome="/home/danis31/Desktop/Temp_docs/reanno_genome"
outpath_genome="/home/danis31/Desktop/Temp_docs/reanno_srna"

reanno <- make_reanno(outpath_genome, PAC=pac_master, mis_fasta_check = TRUE, threads=8)

ls.str(reanno)
names(reanno$Full_anno)
names(reanno$Full_anno$mis0)
colnames(reanno$Full_anno$mis0$genome)

type = "genome"
type = "biotype"

bio_search <- list(
                 Ensembl=c("lncRNA", "miRNA", "pre_miRNA", "rRNA", "snoRNA", "snRNA", "tRNA"),
                 miRNA="dme-",
                 tRNA =c("^tRNA", "mt:tRNA")      # ^ regular expression for start of string  
                 )
                 
new_anno_srna <- add_reanno(anno_srna, bio_search=bio_search, type="biotype", mismatches = 0) 
              
new_anno <- add_reanno(reanno=reanno, mismatches = 3, type="genome")





}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
 documentation about running Sports.
 \url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
 Bowtie \url{https://github.com/Danis102} for updates on the current package.

Other PAC reannotation: 
\code{\link{check_reanno}()},
\code{\link{import_reanno}()},
\code{\link{make_reanno}()},
\code{\link{map_reanno}()},
\code{\link{simplify_reanno}()}
}
\concept{PAC reannotation}
