% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{anno}
\alias{anno}
\alias{anno<-}
\alias{anno,PAC-method}
\title{annotation table}
\usage{
anno(object)

anno(object) <- value

\S4method{anno}{PAC}(object)
}
\description{
annotation table
}
