% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{counts}
\alias{counts}
\alias{counts<-}
\alias{counts,PAC-method}
\title{raw counts table}
\usage{
counts(object)

counts(object) <- value

\S4method{counts}{PAC}(object)
}
\description{
raw counts table
}
