% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{counts}
\alias{counts}
\alias{counts<-}
\alias{counts,PAC-method}
\title{raw counts table}
\usage{
counts(object)

counts() <- value

\S4method{counts}{PAC}(object)
}
\value{
The Counts data frame table from a PAC-object
}
\description{
raw counts table
}
\examples{

load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))

# extra fuctionality with s4 PAC-object:
pac_s4 <- as.PAC(pac)
pac_s4
names(pac_s4)
length(pac_s4)
nrow(pac_s4)
ncol(pac_s4)
rownames(pac_s4)
colnames(pac_s4)
pheno(pac_s4) 
head(anno(pac_s4))
head(counts(pac_s4))
head(norm(pac_s4)$cpm)

pac_s4 <- PAC_summary(pac_s4, norm = "cpm", 
                   type = "means", pheno_target=list("stage")) 
pac_s4                   
head(summary(pac_s4)$cpmMeans_stage)

}
