% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{full}
\alias{full}
\alias{full<-}
\alias{full,reanno-method}
\title{full annotation list of reanno object}
\usage{
full(x)

full(x) <- value

\S4method{full}{reanno}(x)
}
\description{
full annotation list of reanno object
}
