% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_reanno.R
\name{import_reanno}
\alias{import_reanno}
\title{Imports annotation from reannotation mapping}
\usage{
import_reanno(bowtie_path, base = "*.txt", threads = 1,
  report = "biotype")
}
\arguments{
\item{bowtie_path}{Path to a directory where bowtie output files can be
found.}

\item{threads}{Integer stating the number of parallell jobs.}

\item{report}{Character vector "biotype" or "all"}

\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names.}
}
\value{
Data.frame with additional information from reannotation using
  bowtie. If \emph{report="biotype"}, the function will report biotype as the
  file basename of a given reference used in the bowtie reannotation. If
  \emph{report="all"} will give the annotation as full names used in the
  fasta file used as reference in the bowtie reannotation. Caution:
  References with lots of redudancy, such as piRBase fasta files, will
  automatically run with \emph{report="biotype}).
}
\description{
\code{import_reanno} imports bowtie reannotation files.
}
\details{
Given the path to bowtie mapping output, \code{import_reanno} will attempt to
read these files into R and generate a list of unordered data.frames where
each row summarizes all annotations for a given sequence. \code{check_reanno}
will check if sequences are listed in any of the imported reannotation
data.frames. \code{add_reanno} will add the new annotations to a PAC-list
object.
}
\section{Important}{
 Re-annotation must have been done using Bowtie default
  output settings, where output files specifically have been named '.txt'.
  The function will not work with other formats such as SAM/BAM formats.

 The basenames of the bowtie output will also be used to
  annotated what type of annotation. Example: bowtie -v 3 -a -f
  '<piRBase.fasta>' '<master_anno.fasta>' piRNA.txt Will annotate sequences
  as piRNA since the txt-file output was named 'piRNA'
}

\examples{
reanno <- import_reanno(bowtie_path, threads=8)
bowtie_path <- "/data/Data_analysis/Projects/Drosophila/Specific_projects/Test_temp/Processed_Sports_09-10-19/R_files"
bowtie_path <- "/data/Data_analysis/Projects/Drosophila/Specific_projects/Mar_Diet_4_6_2019/Processed_Sports_09-10-19/R_files"
}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports.
  \url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
  Bowtie \url{https://github.com/Danis102} for updates on the current
  package.

Other PAC reannotation: \code{\link{add_reanno}},
  \code{\link{check_reanno}}
}
\concept{PAC reannotation}
