% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_reanno.R
\name{import_reanno}
\alias{import_reanno}
\title{Imports annotation from reannotation mapping}
\usage{
import_reanno(bowtie_path, threads = 1, report = "minimum",
  reduce = "piRNA")
}
\arguments{
\item{bowtie_path}{Path to a directory where bowtie output files can be
found.}

\item{threads}{Integer stating the number of parallell jobs.}

\item{report}{Character vector indicating what to report "minimum" or "full"
(default="minimum")}

\item{reduce}{Character indicating a reference name (without file type
extension) that should be exempted from report="full" and instead will
generate a minimum report.}

\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names.}
}
\value{
List of data frames with additional information from reannotation
  files generated with bowtie -v -a. If \emph{report="minimum"}, the function
  will report each hit only as the number of mismatches for a given reference
  file. If \emph{report="full"} the full annotation reported in the fasta
  file used as reference in the bowtie reannotation will be reported. If a
  reference name is speficied in \emph{reduce}, this reference is excempted
  from \emph{report="full"} and is instead reported as
  \emph{report="minimum"}
  
  Caution: Large references with lots of redudancy (such as pirBase in some
  species) will case massive character strings if \emph{report="full"} with
  no restrictions. Specifing such references in
  \emph{reduce=<reference_names>} will circumvent this problem.
}
\description{
\code{import_reanno} imports bowtie reannotation files.
}
\details{
Given the path to bowtie mapping output, \code{import_reanno} will attempt to
read these files into R and generate a list of unordered data.frames where
each row summarizes all annotations for a given sequence. \code{check_reanno}
will check if sequences are listed in any of the imported reannotation
data.frames. \code{add_reanno} will add the new annotations to a PAC-list
object.
}
\section{Important}{
 Re-annotation must have been done using Bowtie default
  output settings, where output files specifically have been named '.txt'.
  The function will not work with other formats such as SAM/BAM formats.

 The basenames of the bowtie output will also be used to
  annotated what type of annotation. Example: bowtie -v 3 -a -f
  '<piRBase.fasta>' '<master_anno.fasta>' piRNA.txt Will annotate sequences
  as piRNA since the txt-file output was named 'piRNA'
}

\examples{
bowtie_path <- "/data/Data_analysis/Projects/Drosophila/Specific_projects/Mar_Diet_4_6_2019/Processed_Pipeline31_10-03-20/R_files"
bowtie_path <- "/data/Data_analysis/Projects/Drosophila/Other/IOR/Jan_IOR_200130/Data/Single/Processed_Pipeline31_05-03-20/R_files"
reanno1 <- import_reanno(bowtie_path, report="full",  threads=1)
reanno2 <- import_reanno(bowtie_path, report="full",  threads=10)
reanno3 <- import_reanno(bowtie_path, report="full", reduce="piRNA", threads=1)
reanno4 <- import_reanno(bowtie_path, report="full", reduce="piRNA", threads=8)
reanno5 <- import_reanno(bowtie_path, report="minimum")

lapply(reanno5, head)
lapply(reanno5, names)

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports.
  \url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
  Bowtie \url{https://github.com/Danis102} for updates on the current
  package.

Other PAC reannotation: \code{\link{add_reanno}},
  \code{\link{check_reanno}}, \code{\link{make_reanno}},
  \code{\link{simplify_reanno}}
}
\concept{PAC reannotation}
