% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_reanno.R
\name{import_reanno}
\alias{import_reanno}
\title{Imports annotation from reannotation mapping}
\usage{
import_reanno(
  bowtie_path,
  threads = 1,
  coord = FALSE,
  report = "minimum",
  reduce = NULL
)
}
\arguments{
\item{bowtie_path}{Path to a directory where bowtie output files can be
found.}

\item{threads}{Integer stating the number of parallell jobs.}

\item{coord}{Logical whether or not mapping coordinates should be reported
when report="full".}

\item{report}{Character vector indicating what to report "minimum" or "full"
(default="minimum").}

\item{reduce}{Character indicating a reference name (without file type
extension) that should be exempted from report="full" and instead will
generate a minimum report.}

\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names.}
}
\value{
List of data frames with additional information from reannotation
  files generated with bowtie. If \emph{report="minimum"}, the function
  will report each hit only as the number of mismatches for a given reference
  file. If \emph{report="full"} the full name reported in the fasta
  file used as reference in the bowtie reannotation will be reported. If a
  reference name is specified in \emph{reduce}, this reference is excempted
  from \emph{report="full"} and is instead reported as
  \emph{report="minimum"}.
  
  Caution: Large references with lots of redudancy (such as pirBase in some
  species) will cause massive character strings if \emph{report="full"} with
  no restrictions. Specifing such references in
  \emph{reduce=<reference_names>} will circumvent this problem.
}
\description{
This function imports bowtie output and summarizes the content.
}
\details{
Given the path to alignment outputs from bowtie, \code{\link{import_reanno}} will
attempt to read these files into R and generate a list of unordered
data.frames where each row summarizes all annotations for a given sequence.
It is called by \code{\link{map_reanno}} to generate summarized data from bowtie
output saved as .Rdata files. These files can then be re-imported into R
where information is organized using \code{\link{make_reanno}}
and finally added to the annotation table (PAC$Anno) of the original PAC-list
object using \code{\link{add_reanno}}.
}
\section{Important}{
 Re-annotation must have been done using Bowtie default
  output settings, where output files specifically have been named '.txt'.
  The function will not work with other formats such as SAM/BAM formats.

 The basenames of the bowtie output will also be used to
  annotated what type of annotation. Example: bowtie -v 3 -a -f
  '<piRBase.fasta>' '<master_anno.fasta>' piRNA.txt Will annotate sequences
  as piRNA since the txt-file output was named 'piRNA'
}

\examples{
bowtie_path <- "/home/danis31/Desktop/Temp_docs/reanno_genome"
bowtie_path <- "/home/danis31/Desktop/Temp_docs/reanno_srna"
reanno1 <- import_reanno(bowtie_path, report="full",  threads=1)

}
\seealso{
\url{http://bowtie-bio.sourceforge.net/index.shtml} for information
  about Bowtie and for Rbowtie:
  \url{https://www.bioconductor.org/packages/release/bioc/html/Rbowtie.html}.
  \url{https://github.com/Danis102} for updates on the current package.

Other PAC reannotation: 
\code{\link{add_reanno}()},
\code{\link{check_reanno}()},
\code{\link{make_reanno}()},
\code{\link{map_reanno}()},
\code{\link{simplify_reanno}()}
}
\concept{PAC reannotation}
