% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{length,PAC-method}
\alias{length,PAC-method}
\alias{length,reanno-method}
\title{number of objects in PAC}
\usage{
\S4method{length}{PAC}(x)

\S4method{length}{reanno}(x)
}
\arguments{
\item{x}{A S4 object of class PAC.}
}
\value{
The number of items in a PAC-object.

Number of items in the reanno-object.
}
\description{
number of objects in PAC

length of reanno object
}
\examples{

load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))

# extra functionality with s4 PAC-object:
pac_s4 <- as.PAC(pac)
pac_s4
names(pac_s4)
length(pac_s4)
nrow(pac_s4)
ncol(pac_s4)
rownames(pac_s4)
colnames(pac_s4)
pheno(pac_s4) 
head(anno(pac_s4))
head(counts(pac_s4))
head(norm(pac_s4)$cpm)

pac_s4 <- PAC_summary(pac_s4, norm = "cpm", 
                   type = "means", pheno_target=list("stage")) 
pac_s4                   
head(summary(pac_s4)$cpmMeans_stage)

}
