% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_anno.R
\name{make_anno}
\alias{make_anno}
\title{Make a Simple Annotaion File}
\usage{
make_anno(counts)
}
\arguments{
\item{counts}{A Counts list object generated by \code{\link{make_counts}}.}
}
\value{
Simple annotation Data.frame with sequence size (number of characters of
  the Counts list sequences).
}
\description{
\code{make_anno} makes a very simple annotation file
}
\details{
Given a Counts list object generated by \code{\link{make_counts}} this
function extracts sequences lengths and stores them in a data.frame.
}
\examples{
input = system.file("extdata", package = "seqpac", mustWork = TRUE)

counts  <- make_counts(input, threads=6, parse="default_neb",
                       type="fastq", trimming="seqpac", plot=TRUE,
                       evidence=c(experiment=2, sample=1))  
 anno <- make_anno(counts = counts)

}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC generation: 
\code{\link{PAC_check}()},
\code{\link{make_PAC}()},
\code{\link{make_counts}()},
\code{\link{make_cutadapt}()},
\code{\link{make_pheno}()},
\code{\link{make_trim}()},
\code{\link{progress_report}()}
}
\concept{PAC generation}
