% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_counts.R
\name{make_counts}
\alias{make_counts}
\title{Make raw Counts table from Sports output}
\usage{
make_counts(sports_lst, anno, threads)
}
\arguments{
\item{sports_lst}{A list object containing imported data.frames from Sports output.}

\item{anno}{Annotation dataframe.}

\item{threads}{Integer stating the number of parallell jobs.}
}
\value{
The sports_lst list contains output dataframes from Sports and can be
imported to R using \code{import_sports}. This object contains all sequence
counts across all samples.

Using the weakly filtered master annotation file (anno) generated by the
\code{make_anno} function, \code{make_counTable} will extract the target
sequences in anno from sports_lst and merge the counts into one single
data.frame (raw countTable). The output data.frame will ahve rownames
matching the original master annotation file.

Using threads > 1, the final  will be  in parallell Use
\code{parallel::detectcores()} to see available threads on the computer. Do
not exceed the number of available threads!
}
\description{
\code{make_counts} uses parallel processing to generate a raw countTable.
}
\details{
Given a list of imported Sports output data.frames (generated by
\code{import_sports}), and an annotation file (generated by \code{make_anno}
this function will generate a raw countTable (columns=sample, rows=sequence).
}
\examples{
  countTable <- make_countTable(sports_lst = <name_of_your_imported_sports_list>,
                      anno    = <name_of_your_annotation_dataframe>,
                      threads       = <number_of_independent_evidence>,
                      )

  countTable <- make_countTable(sports_lst = sports_lst, anno=anno, threads=12)

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports. \url{https://github.com/Danis102} for
  updates on the current package.

Other PAC generation: 
\code{\link{import_sports}()},
\code{\link{make_PAC}()},
\code{\link{make_anno}()},
\code{\link{make_counts2}()},
\code{\link{make_pheno}()},
\code{\link{progress_report}()}
}
\concept{PAC generation}
