% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_PAC.R
\name{make_PAC}
\alias{make_PAC}
\title{Make PAC}
\usage{
make_PAC(pheno, anno = NULL, counts, output = "S4")
}
\arguments{
\item{pheno}{Data.frame with sample names as row names containing metadata
about each sample. Can be generated by the \code{\link{make_pheno}}
function.}

\item{anno}{Data.frame with unique sequences as row names containing metadata
about each sequence. If anno=NULL, a simple Anno data.frame will
automatically be generated from the the sequence names in counts.
Annotaions for each sequence can be extended using for example the reanno
workflow (see vignette).}

\item{counts}{Data.frame representing a counts table, with column names as
sample names and sequence names as row names. Contains the counts for each
sequence across the samples. Can be generated by the
\code{\link{make_counts}}.}

\item{output}{Character indicating output format. If type="S4" (default),
then the reanno object is returned as an S4 object. If type="S3", the
reanno object will be returned as a list of data.frames}
}
\value{
Ordered PAC list object, checked for compatibility with downstream
  analysis in seqpac.
}
\description{
\code{make_PAC} Compiles, order and checks the assumptions of a PAC file.
}
\details{
Given a pheno, an anno and a counts data.frames the functions will safely
generate a simple PAC list-object.
}
\examples{

library(seqpac) 

### First make counts 


input = system.file("extdata", package = "seqpac", mustWork = TRUE)
counts  <- make_counts(input, threads=1, parse="default_neb",
                       trimming="seqpac", plot=TRUE, 
                       evidence=c(experiment=2, sample=1))


### Then generate a phenotype table with make_pheno (herre using file names)

#  Note:  'Sample_ID' column need to be the same IDs as 
#          colnames in the counts table.

pheno <- as.data.frame(do.call("rbind", 
                       strsplit(list.files(input,pattern="*.fastq.gz"), 
                       "_|\\\\."))[,c(1,2,3,4)]) 
colnames(pheno) <- c("stage", "batch", "index", "sample") 
pheno$Sample_ID <- apply(pheno, 1, function(x){paste(x, collapse="_")}) 
pheno <- make_pheno(pheno=pheno, progress_report=counts$progress_report, 
                    counts=counts$counts)

 
### Lastly combine into PAC

# Note: a simple annotation table will be added automatically.

pac <- make_PAC(pheno=pheno, counts=counts)

}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC generation: 
\code{\link{PAC_check}()},
\code{\link{make_counts}()},
\code{\link{make_cutadapt}()},
\code{\link{make_pheno}()},
\code{\link{make_trim}()},
\code{\link{merge_lanes}()},
\code{\link{progress_report}()}
}
\concept{PAC generation}
