% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_PAC.R
\name{make_PAC}
\alias{make_PAC}
\title{Make PAC}
\usage{
make_PAC(pheno_input, anno_input, counts_input)
}
\arguments{
\item{pheno_input}{Data frame with sample names as row names containing
metadata about each sample. Can be generated by the \code{make_pheno}
function.}

\item{anno_input}{Data frame with unique sequences as row names containing
metadata about each sequence. Can be generated by the \code{make_anno}
function and annotaions for each sequence can be extended using for
example the reanno workflow.}

\item{counts_input}{Data frame with column names as sample names and sequence
names as row names. Contains the counts for each sequence across the
samples. Can be generated by the \code{make_counts} \code{make_counts2}
function.}
}
\value{
Ordered and assumption checked PAC list object, that can be used in
  downstream analysis.
}
\description{
\code{make_PAC} Compiles, order and checks the assumptions of a PAC file.
}
\details{
Given a pheno, an anno and a counts data frames the functions will safely
generate simple PAC list-object.
}
\examples{

### Load counts
load(file="/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/temp.Rdata")

## Make Anno
anno <- make_anno(input=counts, type="counts")


## Make pheno
pheno_input <- read.delim("/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/Groups_data.txt", sep=",", header=TRUE)
pheno_input$Sample_ID <- gsub("_s", "_notag",  as.character(pheno_input$Sample_ID))
pheno_input$Sample_ID <- gsub("_d", "_tag",  as.character(pheno_input$Sample_ID))

pheno <- make_pheno(pheno_input=pheno_input, type="manual", countTable=counts[[1]])

pheno_input=pheno
anno_input=anno
counts_input=counts[[1]]

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports. \url{https://github.com/Danis102} for
  updates on the current package.

Other PAC generation: \code{\link{import_sports}},
  \code{\link{make_anno}}, \code{\link{make_counts2}},
  \code{\link{make_counts}}, \code{\link{make_pheno}},
  \code{\link{progress_report}}
}
\concept{PAC generation}
