% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pheno.R
\name{make_pheno}
\alias{make_pheno}
\title{Make Phenotype File}
\usage{
make_pheno(pheno_file_path, countTable = NULL, progress_report = NULL)
}
\arguments{
\item{pheno_file_path}{Character vector with the path to a .csv file.}

\item{countTable}{Data.frame object with the same column names as in Sample_ID column of the .csv file.}

\item{countTable}{Data.frame object with progress report.}
}
\value{
data.frame
}
\description{
\code{make_pheno} generates a Pheno data.frame object containing additional
metadata for unique samples.
}
\details{
Given the path to a directory with a single csv file, the function will read
the file. If provided with a countTable (see \code{make_countTable}), it will
attempt to organize the the row names according to the column names in the
countTable.

The function was specifically written for Illumina SampleSheetUsed.csv files
downloaded from the BaseSpace server using the basemount software. It will,
however, be compatible with any comma seperated text file with any of its
first lines containing a header with the first column named 'Sample_ID'.

The function will also attempt to read a 'Samples_stat.txt' (tab seperate
file) file also downloaded from Illumina BaseSpace. This file will add
original number of reads (past filter) generated in the sequencing. If
provided with a Sports progress_report object (see \code{progress_report} the
script will attempt to add this data to the Pheno data.frame.
}
\examples{
report <- make_pheno(<your_path_to_sports_output_directory>)
pheno <- make_pheno(pheno_file_path, countTable=countTable, progress_report=report)

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports. \url{https://github.com/Danis102} for
  updates on the current package.

Other PAC generation: \code{\link{import_sports}},
  \code{\link{make_anno}}, \code{\link{make_countTable}},
  \code{\link{progress_report}}
}
\concept{PAC generation}
