% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pheno.R
\name{make_pheno}
\alias{make_pheno}
\title{Make Phenotype File}
\usage{
make_pheno(pheno, type = "manual", counts = NULL, progress_report = NULL)
}
\arguments{
\item{pheno}{Character vector with the path to a .csv file or a
data.frame with a column named "Sample_ID".}

\item{type}{Character indicating what type to file to expect. If
type="manual", the function will attempt to read a file named 'pheno.csv',
in which the a column has been named 'Sample_ID' containing the exact
sample names matching basenames of the fastq sample files. If
\emph{type="basespace"} the function will attempt to read and join
information from 'SampleSheetUsed.csv' and 'Samples_stat.txt' in
pheno that have been downloaded from Illumina Basespace.}

\item{counts}{Data.frame object with the same column names as in
Sample_ID column of the .csv file.}

\item{progress_report}{Data.frame object with progress report.}
}
\value{
data.frame
}
\description{
\code{make_pheno} generates a Pheno data.frame object containing additional
metadata for unique samples.
}
\details{
Given the path to a directory with a single csv file, the function will read
the file. Alternatively, a data.frame with sample information can be
provided. If provided with a count table, (see \code{\link{make_counts}}), it
will attempt to organize the the row names according to the column names in
the counts table.

The function was originally written for Illumina SampleSheetUsed.csv files
downloaded from the BaseSpace server using the basemount software. It will,
however, be compatible with any comma seperated text file with any of its
first lines containing a header with the first column named 'Sample_ID'.

The function will also attempt to read a 'Samples_stat.txt' (tab seperate
file) file also downloaded from Illumina BaseSpace. This file will add
original number of reads (past filter) generated in the sequencing. If
provided with a progress_report object (see \code{\link{make_counts}} the
script will attempt to add this data to the Pheno data.frame.
}
\examples{

library(seqpac) 

### First make counts 

input = system.file("extdata", package = "seqpac", mustWork = TRUE)
counts  <- make_counts(input, threads=1, parse="default_neb", type="fastq",
                       trimming="seqpac", plot=TRUE, evidence=c(experiment=2, sample=1))


### Then generate a phenotype table with make_pheno (herre using file names)

#  Note:  'Sample_ID' column need to be the same IDs as colnames in the counts table.

pheno <- as.data.frame(do.call("rbind", strsplit(list.files(input,pattern="*.fastq.gz"), "_|\\\\."))[,c(1,2,3,4)]) 
colnames(pheno) <- c("stage", "batch", "index", "sample") 
pheno$Sample_ID <- apply(pheno, 1, function(x){paste(x, collapse="_")}) 
pheno <- make_pheno(pheno=pheno, progress_report=counts$progress_report, counts=counts$counts)

 
### Lastly combine into PAC

# Note: a simple annotation table will be added automatically.

pac <- make_PAC(pheno=pheno, counts=counts, anno=NULL)

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports. \url{https://github.com/Danis102} for
  updates on the current package.

Other PAC generation: 
\code{\link{PAC_check}()},
\code{\link{make_PAC}()},
\code{\link{make_counts}()},
\code{\link{make_cutadapt}()},
\code{\link{make_trim}()},
\code{\link{progress_report}()}
}
\concept{PAC generation}
