% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reanno.R
\name{make_reanno}
\alias{make_reanno}
\title{Makes annotation from R imported reannotation mapping}
\usage{
make_reanno(reanno_path, PAC, mis_fasta_check = FALSE, threads = 1)
}
\arguments{
\item{reanno_path}{Path to a directory where reannotation .Rdata files can be
found.}

\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names.}

\item{mis_fasta_check}{Logical TRUE/FALSE if checking against anno_misX.fa
should be done availble in the same folder as the reannoration files.}

\item{threads}{Integer indicating the number of paralell workers to be used.}
}
\value{
List of dataframes with additional information from reannotation
  using bowtie. If \emph{mis_fasta_check} is specified the function will look
  for a \emph{anno_misX.fa} (X = the file with the highest number) previoulsy
  generated by the reannotation workflow. This file is used to double check
  so that no sequences are missing before and after reannotation.
}
\description{
\code{make_reanno} makes a reannotation list.
}
\details{
Given the path to the imported R reannotation files
(reanno_mis0/1/2/3/4/5.Rdata) generated by import_reanno, this function will
summarize the reannotation files into one output that matches the order of
sequences in a PAC object.
}
\examples{
reanno_path  <- "/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/reanno"

load(file="/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/PAC_hifilt.Rdata")

Full_anno <- make_reanno(reanno_path, PAC=PAC_hifilt, mis_fasta_check=TRUE, threads=10)    # Complete use
identical(rownames(PAC_master$Anno), rownames(Full_anno$Overview)) 


}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports.
  \url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
  Bowtie \url{https://github.com/Danis102} for updates on the current
  package.

Other PAC reannotation: 
\code{\link{add_reanno}()},
\code{\link{check_reanno}()},
\code{\link{import_reanno}()},
\code{\link{map_reanno}()},
\code{\link{simplify_reanno}()}
}
\concept{PAC reannotation}
