% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reanno.R
\name{map_reanno}
\alias{map_reanno}
\title{Makes annotation from R imported reannotation mapping}
\usage{
map_reanno(
  PAC,
  type = "bowtie",
  output_path,
  ref_paths,
  mismatches = 3,
  mis_fasta_check = FALSE,
  threads = 1,
  par_type = "PSOCK",
  parse_bowtie = "bowtie -a --norc -f"
)
}
\arguments{
\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names.}

\item{type}{Character indicating if mapping should be performed by bowtie or}

\item{mis_fasta_check}{Logical TRUE/FALSE if checking against anno_misX.fa
should be done availble in the same folder as the reannoration files.}

\item{threads}{Integer indicating the number of paralell workers to be used.}
}
\value{
List of dataframes with additional information from reannotation
  using bowtie. If \emph{mis_fasta_check} is specified the function will look
  for a \emph{anno_misX.fa} (X = the file with the highest number) previoulsy
  generated by the reannotation workflow. This file is used to double check
  so that no sequences are missing before and after reannotation.
}
\description{
\code{map_reanno} makes a reannotation list.
}
\details{
Given the path to the imported R reannotation files
(reanno_mis0/1/2/3/4/5.Rdata) generated by import_reanno, this function will
summarize the reannotation files into one output that matches the order of
sequences in a PAC object.
}
\examples{
load(file="/home/danis31/OneDrive/Programmering/Programmering/Pipelines/Drosophila/Pipeline_3.1/seqpac/dm_test_PAC.Rdata")

#' #PAC_filt <- PAC_filter(PAC_all, threshold=5, coverage=4, type="counts", stat=FALSE, pheno_target=NULL, anno_target=NULL)


ref_paths <- list(miRNA="/data/Data_analysis/Genomes/Drosophila/dm6/sports/Drosophila_melanogaster/miRNA/miRBase_21-dme.fa",
                  Ensembl="/data/Data_analysis/Genomes/Drosophila/dm6/sports/Drosophila_melanogaster/Ensembl/Drosophila_melanogaster.BDGP6.ncrna.fa",
                  rRNA="/data/Data_analysis/Genomes/Drosophila/dm6/sports/Drosophila_melanogaster/rRNA_reanno/drosophila_rRNA_all.fa",
                  tRNA="/data/Data_analysis/Genomes/Drosophila/dm6/sports/Drosophila_melanogaster/tRNA_reanno/tRNA_mature.fa",
                  piRNA="/data/Data_analysis/Genomes/Drosophila/dm6/sports/Drosophila_melanogaster/piRNA_piRBase/piR_dme.fa")


PAC = PAC_filt
output_path= "/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/reanno/"
mismatches=3
threads=10
par_type="PSOCK"
parse_bowtie <- "bowtie -a --norc -f"   

reanno <- map_reanno(PAC_filt, output_path= output_path, ref_paths=ref_paths,  mismatches=3, threads=10)

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports.
  \url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
  Bowtie \url{https://github.com/Danis102} for updates on the current
  package.

Other PAC reannotation: 
\code{\link{add_reanno}()},
\code{\link{check_reanno}()},
\code{\link{import_reanno}()},
\code{\link{make_reanno}()},
\code{\link{simplify_reanno}()}
}
\concept{PAC reannotation}
