% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lanes.R
\name{merge_lanes}
\alias{merge_lanes}
\title{Merge fastq lane files}
\usage{
merge_lanes(in_path, out_path, threads = 1)
}
\arguments{
\item{in_path}{Character string with the path to lane files that should be
merged. The names of the lane files should follow the convention:
'sample1_lane1.fastq.gz, sample1_lane2.fastq.gz, sample2_lane1.fastq.gz,
etc', where the first part of the name indicate the sample while the second
part indicate lanes. The function will automatically identify samples by
file names with idenitcal first parts. The function only takes fastq.gz
compressed files.}

\item{out_path}{Character string with the path to destination folder for
merged fastq.gz files.}

\item{threads}{Integer indicating the number of parallel processes that
should be used.}
}
\value{
Merged fastq files in destination folder.
}
\description{
\code{merge_lanes} Identifies, reads and merges fastq lane files.
}
\details{
Given an input path were flow cell lane files in fastq format (typically
generated by for example Illumina sequencers) this function will try to
identify unique samples among the lane files and merge all lane files for
each sample. The function uses md5 checks to control that each lane file is
unique (no accidental copies) and will compare expected outcomes given that
all samples should have the same number of lanes (comming from the same
experiment).
}
\examples{

\dontrun{
in_path <- "/some/path/to/fastq.gz/lane/files"
out_path <- "/some/path/to/fastq.gz/merged/files"

merge_lanes(in_path, out_path, threads=12)

}
}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC generation: 
\code{\link{PAC_check}()},
\code{\link{make_PAC}()},
\code{\link{make_counts}()},
\code{\link{make_cutadapt}()},
\code{\link{make_pheno}()},
\code{\link{make_trim}()},
\code{\link{progress_report}()}
}
\concept{PAC generation}
