% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{names,PAC-method}
\alias{names,PAC-method}
\alias{rownames,PAC-method}
\alias{names,reanno-method}
\title{sequence names}
\usage{
\S4method{names}{PAC}(x)

\S4method{rownames}{PAC}(x)

\S4method{names}{reanno}(x)
}
\description{
sequence names

names of objects in reanno
}
