% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{norm}
\alias{norm}
\alias{norm<-}
\alias{norm,PAC-method}
\title{list of normalized tables}
\usage{
norm(object)

norm(object) <- value

\S4method{norm}{PAC}(object)
}
\description{
list of normalized tables
}
