% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{nrow,PAC-method}
\alias{nrow,PAC-method}
\alias{nrow,reanno-method}
\title{number of sequences}
\usage{
\S4method{nrow}{PAC}(x)

\S4method{nrow}{reanno}(x)
}
\arguments{
\item{x}{A S4 object of class PAC.}
}
\value{
The number of sequences in a PAC-object.

Number of sequences in the reanno-object.
}
\description{
number of sequences

number of sequences
}
\examples{

load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))

# extra fuctionality with s4 PAC-object:
names(pac)
length(pac)
nrow(pac)
ncol(pac)
rownames(pac)
colnames(pac)
pheno(pac) 
head(anno(pac))
head(counts(pac))
head(norm(pac)$cpm)

}
