% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{nrow,PAC-method}
\alias{nrow,PAC-method}
\alias{nrow,reanno-method}
\title{number of sequences}
\usage{
\S4method{nrow}{PAC}(x)

\S4method{nrow}{reanno}(x)
}
\description{
number of sequences

number of sequences
}
