% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{overview}
\alias{overview}
\alias{overview<-}
\alias{overview,reanno-method}
\title{overview table of reanno object}
\usage{
overview(x)

overview(x) <- value

\S4method{overview}{reanno}(x)
}
\description{
overview table of reanno object
}
