% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_check.R
\name{PAC_check}
\alias{PAC_check}
\title{Checks PAC format}
\usage{
PAC_check(PAC, sample_check = NULL, seq_check = NULL)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{sample_check}{Can be provided either as a data frame or the name of a
data frame in PAC. Will check both the row and column names of the object
for identical sample names in PAC. A simple character vector can also be
provided.}

\item{seq_check}{Can be provided either as a data frame or the name of an
data frame in PAC. Will check only the row names of the object for
identical sequence names in PAC. A simple character vector can also be
provided.}
}
\value{
If PAC format is correct will return logical TRUE, else it will throw an error.
}
\description{
\code{PAC_check} Quickly checks rownames in PAC.
}
\details{
Given a PAC object the function will cross check the rownames and colnames
across Pheno, Anno and Counts.
}
\examples{

load("~/OneDrive/Programmering/Programmering/Pipelines/Drosophila/Pipeline_3.1/seqpac/dm_test_PAC.Rdata")

PAC_filt$summary$pheno <- PAC_filt$Pheno
PAC_filt$summary$pheno_trans <- t(PAC_filt$Pheno)
PAC_filt$summary$pheno_trans2 <- t(PAC_filt$Pheno)[,-1]
PAC_filt$summary$pheno_rev <- PAC_filt$Pheno[rev(seq(PAC_filt$Pheno)),]


PAC_check(PAC_filt, sample_check="pheno") # TRUE
PAC_check(PAC_filt, sample_check="pheno_trans") # TRUE (Automatically looking for samples in columns instead)
PAC_check(PAC_filt, sample_check="pheno_trans2") #Error with hint
PAC_check(PAC_filt, sample_check="pheno_rev") #Error not identical

PAC_filt$summary$anno <- PAC_filt$Anno
PAC_filt$summary$anno_trans <- t(PAC_filt$Anno)
PAC_filt$summary$anno_mis <- PAC_filt$Anno[-567,]

PAC_check(PAC_filt, seq_check="anno") # TRUE
PAC_check(PAC_filt, seq_check="anno_trans") # Error not identical
PAC_check(PAC_filt, seq_check="anno_mis") # Error not identical

PAC_check(PAC_filt, seq_check=PAC_filt$summary$anno) #TRUE
PAC_check(PAC_filt, seq_check=PAC_filt$summary$anno_trans)  # Error not identical
PAC_check(PAC_filt, seq_check=PAC_filt$summary$pheno) #Error not identical
PAC_check(PAC_filt, sample_check=PAC_filt$summary$pheno_rev) # Error not identical
PAC_check(PAC_filt, sample_check=PAC_filt$summary$pheno_trans) # TRUE (Automatically looking for samples in columns instead)
PAC_check(PAC_filt, sample_check=rownames(PAC_filt$summary$pheno)) # TRUE
PAC_check(PAC_filt, sample_check=rownames(PAC_filt$summary$pheno_trans)) # Error not identical

PAC_check(PAC_filt, sample_check=as.integer(13))
PAC_check(PAC_filt, sample_check=list(13,45))

}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.
}
\concept{}
