% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_check.R
\name{PAC_check}
\alias{PAC_check}
\title{Checks PAC format}
\usage{
PAC_check(PAC, sample_check = NULL, seq_check = NULL)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{sample_check}{Can be provided either as a data frame or the name of a
data frame in PAC. Will check both the row and column names of the object
for identical sample names in PAC. A simple character vector can also be
provided.}

\item{seq_check}{Can be provided either as a data frame or the name of an
data frame in PAC. Will check only the row names of the object for
identical sequence names in PAC. A simple character vector can also be
provided.}
}
\value{
If PAC format is correct will return logical TRUE, else it will throw an error.
}
\description{
\code{PAC_check} Quickly checks rownames in PAC.
}
\details{
Given a PAC object the function will cross check the rownames and colnames
across Pheno, Anno and Counts.
}
\examples{

\dontrun{
library(seqpac)
load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))

PAC_check(pac)  # TRUE

# Remove first row in Anno
new_Anno <- pac$Anno[-1,]               


#PAC_check(pac, seq_check=new_Anno) # Error   

# Add to pac
pac$Anno <- new_Anno                    
#PAC_check(pac) # Error                      

## Reload data
load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))
PAC_check(pac)  # TRUE

# Remove a sample column in normalized table 
new_norm <- pac$norm$cpm[,-1]           
PAC_check(pac, seq_check=new_norm)      # All sequences are good
#PAC_check(pac, sample_check=new_norm)  # Error

pac$norm$cpm <- new_norm                # Add to pac
#PAC_check(pac)                         # Error 

## Reload data
load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))
PAC_check(pac)  # TRUE

seq_names <- rownames(pac$Counts)
sample_names <- colnames(pac$Counts) 

PAC_check(pac, seq_check=seq_names)
PAC_check(pac, sample_check=sample_names) 

#PAC_check(pac, seq_check=rev(seq_names))
#PAC_check(pac, sample_check=rev(sample_names)) 
}

}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC generation: 
\code{\link{make_PAC}()},
\code{\link{make_counts}()},
\code{\link{make_cutadapt}()},
\code{\link{make_pheno}()},
\code{\link{make_trim}()},
\code{\link{merge_lanes}()},
\code{\link{progress_report}()}
}
\concept{PAC generation}
