% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_covplot.R
\name{PAC_covplot}
\alias{PAC_covplot}
\title{Plot sequence covarage over a reference}
\usage{
PAC_covplot(
  PAC,
  map,
  summary_target = names(PAC),
  map_target = NULL,
  style = "line",
  xseq = TRUE,
  color = c("black", "red", "grey", "blue"),
  check_overide = FALSE
)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{summary_target}{List with 2 character vectors: 
1st object: Character indicating the name of the target PAC summary object
which should be used as input data for the plots. If left empty the 1st
summary object will be used.
2nd object: Character vector indicating the names of the columns in the
summary object from which individual graphs should be plotted. If left
empty all summary columns are plotted in the same graph.
Summaries are generated by the PAC_summary function or customly generated
by the user and stored in the summary 'folder' of a PAC object.}

\item{map_target}{(optional) Character vector. Imporant: This is similar to
an anno_target, but instead extracts target references in the PAC_mapping
object. Should contain search terms that can find unique strings in the
reference names. The search terms are parse to grepl("<search terms>,
names(<PAC_mapping object>)). (default=NULL)}

\item{xseq}{Logical indicating whether the nucleotides should be plotted on
the x-axis. Plotting long references with xseq=FALSE will increase
script performance. (default=TRUE).}

\item{color}{Character vector indicating the rgb colors to be parsed to
ggplot2 for plotting the covarage lines  (default = c("black", "red",
"grey", "blue"))}

\item{check_overide}{Logical, whether all sequences names in PAC must be
found in the PAC_map object (TRUE) or not (FALSE). If the PAC_map object
was generated from the provided PAC objects, then all sequences should
match. If the PAC list object was subsetted or the PAC_map is used with an
independent PAC list object in which not all sequences are available, then
this function will create an error, unless check_overide=TRUE. In such
case, all missing sequences in the PAC object will automatically be
assigned a 0-value in all summary_target columns.}

\item{PAC_map}{PAC_map object generated by \emph{PAC_mapping}.}
}
\value{
Plot list with plots generated by ggplot2
}
\description{
\code{PAC_covplot} Plotting sequences in a PAC object using an PAC mapping object.
}
\details{
Given a PAC object and a PAC_map object generated by \emph{PAC_mapping}
this function will attempt to plot the sequence coverage over long
references.
}
\examples{

### Load data ###
path="/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/"
load(file=paste0(path, "PAC_all.Rdata"))

PAC_filt <- PAC_filter(PAC_all, size=c(16,70), threshold=10, coverage=5, type="counts", stat=FALSE, pheno_target=NULL, anno_target=NULL)
PAC_filt <- PAC_rpm(PAC_filt)
PAC_filt <- PAC_filter(PAC_filt, size=c(16,70), threshold=10, coverage=5, type="rpm", stat=FALSE, pheno_target=NULL, anno_target=NULL)

## Remove corrupt samples and make means 
Ph_trg <- as.character(PAC_filt$Pheno$Sample[!PAC_filt$Pheno$Sample \%in\% "Inx24_200130_S12"])
PAC_filt <- PAC_filter(PAC_filt, pheno_target= list("Sample", Ph_trg))

PAC_filt$Pheno$Groups <- paste(do.call("rbind", strsplit(as.character(PAC_filt$Pheno$SampleProject), "_" ))[,1], PAC_filt$Pheno$Method, PAC_filt$Pheno$Method_tag, PAC_filt$Pheno$Tag, sep="_")

## Make summaries
PAC_filt <- PAC_summary(PAC_filt, norm = "rpm", type = "means", pheno_target=list("Groups", unique(PAC_filt$Pheno$Groups)))

## Mapping
map_rRNA <- PAC_mapper(PAC_filt, ref_path="/data/Data_analysis/Genomes/Drosophila/dm6/sports/Drosophila_melanogaster/rRNA_reanno/drosophila_rRNA_all.fa", threads=12)

All_plots <- lapply(as.list(Smry_trg_all), function(x){PAC_covplot(PAC_filt, map_rRNA, summary_target = list("means_[Groups]", x), xseq=FALSE, style="line", color="red")})

cowplot::plot_grid(All_plots[[1]][[7]], All_plots[[2]][[7]], All_plots[[3]][[7]], 
                  All_plots[[4]][[7]], All_plots[[5]][[7]], All_plots[[6]][[7]],
                  All_plots[[7]][[7]], All_plots[[8]][[7]], All_plots[[9]][[7]], 
                  All_plots[[10]][[7]], All_plots[[11]][[7]], All_plots[[12]][[7]],
                  All_plots[[13]][[7]], All_plots[[14]][[7]], All_plots[[15]][[7]],
                  nrow = 5, ncol = 3)




}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_rpm}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_tRNA}()},
\code{\link{map_rangetype}()}
}
\concept{PAC analysis}
