% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_filtsep.R
\name{PAC_filtsep}
\alias{PAC_filtsep}
\title{Generates filtered sequence name list seperated on group}
\usage{
PAC_filtsep(
  PAC,
  type = "counts",
  threshold = 10,
  coverage = 100,
  pheno_target = NULL
)
}
\arguments{
\item{PAC}{PAC-list object containing an Pheno data.frame with samples as row
names and a Counts table with raw counts or normalized object containing
for example reads per million (rpm; generated by \code{PAC_rpm}).

PAC, norm="raw", threshold=10, coverage=100, pheno_target=NULL}

\item{type}{Character specifying if filtering should be done using "counts",
"rpm" or another normalized data table in PAC$norm (default="counts").}

\item{coverage}{Integer giving the percent of independent samples of each
group that need to reach the threshold for a sequence to be included
(default=100).}

\item{pheno_target}{(optional) List with: 
1st object being a character vector of target column in Pheno, 
2nd object being a character vector of the target group(s) in the target Pheno column (1st object).
(default=NULL)}

\item{treshold}{Integer giving the threshold in rpm or counts that needs to
be reached for a sequence to be included (default=10).}
}
\value{
A data frame where each column contains the sequences that passed the
  filter for a specific group specified in pheno_target.
}
\description{
\code{PAC_filtsep} Group seperated filtered name list
}
\details{
Given a PAC object the function will extract sequences within the reach a
given filter within a given group.
}
\examples{

}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_gtf}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_rpm}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_tRNA}()},
\code{\link{map_rangetype}()}
}
\concept{PAC analysis}
