% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_map.R
\name{PAC_map}
\alias{PAC_map}
\title{Aligning sequences against references to produce annotation information}
\usage{
PAC_map(
  PAC,
  type = "internal",
  mapping = "biotype",
  output,
  input,
  bio_search = NULL,
  mismatches = 3,
  save_name = FALSE,
  threads = 1,
  override = FALSE,
  ...
)
}
\arguments{
\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names.}

\item{type}{Character indicating if mapping should be performed by calling
the internal bowtie function of the Rbowtie package (type="internal"), or
if bowtie should be called externally through a system call for a locally
installed bowtie (type="external").}

\item{mapping}{Character or a list. If \code{mapping="genome"} mapping is done
against a reference genome and genomic coordinates are acquired. If
mapping="biotype", mapping is done against a specialized fasta reference
(e.g. Ensembl_ncrna, pirBase etc), where genomic coordinates is not
required because classification will be performed on a match-or-no-match
basis. A list of exactly 3 objects, named "coord", "report" and "reduce"
can also be provided. This list will be parsed to the
\code{\link{import_reanno}} function. When mapping="genome", the list
\code{import=list(coord=TRUE, report="full", reduce=NULL)} is automatically
parsed, while when mapping="biotype" the list parsed is
\code{import=list(coord=FALSE, report="full", reduce=NULL)}. Performance
increases by setting coord=FALSE. See \code{\link{import_reanno}} for more
information on how to set \code{report} and \code{reduce} for increased
performance when extremely large and repetitive references are used, such
as pirBase and repeatMasker.}

\item{output}{Character indicating the path to the destination folder
for the files generated by the function. If no output folder is defined,
files will be stored in a temporary folder and may be removed on system
restart.}

\item{input}{List of file paths (character) indicating the full paths to
each fasta reference file. It is important to carefully name each reference path.
The names will appear as they are named in the final annotation table.
Thus, if \emph{input=list(tRNA="<path_tRNA_ref>",
miRNA="<path_miRNA_ref>")} mapping to these fasta references will appear as
"tRNA" and "miRNA", respectively. Note: All reference fasta files must have
bowtie indexes using \code{Rbowtie::bowtie_build}.}

\item{bio_search}{List of character vectors indicating search terms targeting
feature names in the original reference fasta files. These search terms will
be parsed to \code{\link{grepl}} as regular expressions. List names must
match names of the references in the reanno object. Classifications will be
reported with reference name + search term (e.g. "Ensembl_trna"). This parameter
works with \code{\link{add_reanno}}, please see that function for more detailed 
information.}

\item{mismatches}{Integer indicating the maximum number of mismatches allowed
in the alignments. The function currently supports no more than 3
mismatches (botwie max).}

\item{save_name}{logical statement whether to report back full length names of
biotype annotations. The default is set to FALSE, as this may create Anno objects
of wonky dimensions with certain reference .fasta files.}

\item{threads}{Integer indicating the number of parallel processes to be
used.}

\item{override}{logical whether to write over files in output folder (TRUE),
or whether to ask user for}

\item{...}{Arguments to be passed on to \code{\link{map_reanno}}, \code{\link{make_reanno}}
and \code{\link{add_reanno}}.}
}
\value{
Will return a PAC object with an expanded Anno.
}
\description{
Aligns sequences in a PAC list object against specified references, returning
a data frame or PAC object containing annotation information
}
\details{
Given a PAC object, the destination of reference files, and information of
what type of information to save, this function will perform mapping against references
and retain information of mapping. This function is based of three main functions;
 \code{map_reanno}, \code{make_reanno} and \code{add_reanno}, whom may be ran 
 separately for more detailed control of the workflow. 
 Using save_name=TRUE in this function will retain the name of the references where sequences are found,
 which may cause large objects.
}
\examples{

# Example for general mapping, for more advanced uses please see
# ?map_reanno, ?simply_reanno an ?add_reanno
closeAllConnections()

# Create an reanno object

## First, if you haven't already generated Bowtie indexes for the included
# fasta references you need to do so. If you are having problem see the small
# RNA guide (vignette) for more info.
                                                             
 ## tRNA:
 trna_file <- system.file("extdata/trna", "tRNA.fa", 
                          package = "seqpac", mustWork = TRUE)
 trna_dir <- dirname(trna_file)
 
 if(!sum(stringr::str_count(list.files(trna_dir), ".ebwt")) ==6){     
     Rbowtie::bowtie_build(trna_file, 
                           outdir=trna_dir, 
                           prefix="tRNA", force=TRUE)
                           }
 ## rRNA:
 rrna_file <- system.file("extdata/rrna", "rRNA.fa", 
                          package = "seqpac", mustWork = TRUE)
 rrna_dir <- dirname(rrna_file)
 
 if(!sum(stringr::str_count(list.files(rrna_dir), ".ebwt")) ==6){
     Rbowtie::bowtie_build(rrna_file, 
                           outdir=rrna_dir, 
                           prefix="rRNA", force=TRUE)
                           }
                           
                           
##  Then load a PAC-object and remove previous mapping from anno:
 load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                   package = "seqpac", mustWork = TRUE))
 anno(pac) <- anno(pac)[,1, drop = FALSE]
 
 input <- list(trna= trna_file, rrna= rrna_file)

                                    
##  You may add an output path of your choice, but here we use a temp folder:
 output <- paste0(tempdir(),"/seqpac/test")

bio_search <- list(rrna=c("5S", "5.8S", "12S", "16S", "18S", "28S", "pre_S45"),
trna =c("_tRNA", "mt:tRNA"))

pac_anno <- PAC_map(PAC=pac, output=output,
mismatches=0,
input=input, mapping="biotype", save_name=TRUE,
bio_search=bio_search, override=TRUE)

head(pac_anno)

}
\seealso{
\url{http://bowtie-bio.sourceforge.net/index.shtml} for information
  about Bowtie and for Rbowtie:
  \url{https://www.bioconductor.org/packages/release/bioc/html/Rbowtie.html}.
  \url{https://github.com/OestLab/seqpac} for updates on the current package.

Other PAC reannotation:
\code{\link[=PAC_annotate]{PAC_annotate()}},
\code{\link[=PAC_map2]{PAC_map2()}},
\code{\link[=add_reanno]{add_reanno()}},
\code{\link[=as.reanno]{as.reanno()}},
\code{\link[=import_reanno]{import_reanno()}},
\code{\link[=make_conv]{make_conv()}},
\code{\link[=make_reanno]{make_reanno()}},
\code{\link[=map_reanno]{map_reanno()}},
\code{\link[=simplify_reanno]{simplify_reanno()}}
}
\concept{PAC reannotation}
