% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_mapper.R
\name{PAC_mapper}
\alias{PAC_mapper}
\title{Advanced sequences mapping of a PAC object}
\usage{
PAC_mapper(PAC, ref_path, mismatches = 0, threads = 1,
  par_type = "PSOCK")
}
\arguments{
\item{PAC}{PAC-list object.}

\item{ref_path}{Character indicating the path to the fasta (.fa) reference
file.}

\item{mismatches}{Integer indicating the number of mismatches that should be
allowed in the mapping.}

\item{threads}{Integer indicating the number of parallel processes that
should be used.}

\item{par_type}{Character indicating if "PSOCK" or "FORK" should be parsed to
parallel::makeCluster (default="PSOCK").}

\item{mismatches=0}{}
}
\value{
Stacked list, where each object on the highest level contains:
                   (Object 1) Reference name and sequence. 
                   (Object 2) Dataframe showing the mapping results of
                              each quiary sequence that mapped to Object 1.
}
\description{
\code{PAC_mapper} Mapping sequences against a reference.
}
\details{
Given a PAC object and the path to a fasta reference file, this function will
map sequences in PAC and mapping extract information.
}
\examples{

### Load data ###
reanno_path="/data/Data_analysis/Projects/Pigs/Specific_projects/SRA_download/SRP135969_Sperm_Exosomes_Hemicastration/Processed_Pipeline31_05-03-20/R_files/"
load(file=paste0(reanno_path, "PAC_filt_rpm10in25.Rdata"))



table(PAC_filt$Anno$Biotypes_tRNA)
PAC_filt_tRNA <- PAC_filter(PAC_filt, anno_target=list("Biotypes_tRNA", c("Mt_tRNA", "tRNA")), subset_only=TRUE)
table(PAC_filt_tRNA$Anno$Biotypes_tRNA)

ref_path <- "/data/Data_analysis/Genomes/Pigs/Sports/Sus_scrofa/tRNA_reanno/tRNA_mature.fa"
ref_path <- "/data/Data_analysis/Genomes/Pigs/Sports/Sus_scrofa/tRNA/tRNA_mature.fa"
PAC_map <- PAC_mapper(PAC=PAC_filt_tRNA, ref_path, threads=8, mismatches=0)

table(rownames(PAC_filt_tRNA$Anno) \%in\%  unique(do.call("c", lapply(PAC_map, function(x){ rownames(x[[2]])}))))

save(extended_anno_lst, file=paste0(reanno_path, "/Extended_anno_tsRNA.Rdata"))


}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: \code{\link{PAC_covplot}},
  \code{\link{PAC_deseq}}, \code{\link{PAC_filter}},
  \code{\link{PAC_jitter}}, \code{\link{PAC_nbias}},
  \code{\link{PAC_pca}}, \code{\link{PAC_pie}},
  \code{\link{PAC_rpm}}, \code{\link{PAC_saturation}},
  \code{\link{PAC_stackbar}}, \code{\link{PAC_summary}}
}
\concept{PAC analysis}
