% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_mapper.R
\name{PAC_mapper}
\alias{PAC_mapper}
\title{Advanced sequence mapping of a PAC object}
\usage{
PAC_mapper(
  PAC,
  ref,
  mapper = "reanno",
  mismatches = 0,
  threads = 1,
  N_up = "",
  N_down = "",
  report_string = FALSE
)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{ref}{Character indicating the path to the fasta (.fa) reference
file or a DNAStringSet with already loaded reference sequences.}

\item{mapper}{Character indicating what align engine to be used. If
mapper="reanno" (default) PAC_mapper will create temporary files and
generate missing bowtie indexes for a reannotion workflow (see
\code{?map_reanno}. Important, for PAC_mapper to successfully tap into the
reannotation workflow, the names of ref (fasta or DNAStringSet) cannot
contain any ";" characters. If mapper="vmatch", PAC_mapper will instead use
the \code{vmatchPattern} function in the Biostrings package. While the
mapper="reanno" is by far the fastest option, mapper="vmatch" does not
require bowtie indexes.}

\item{mismatches}{Integer indicating the number of mismatches that should be
allowed in the mapping.}

\item{threads}{Integer indicating the number of parallel processes that
should be used.}

\item{N_up}{Character indicating a sequence that should be added to the
reference at the 5' end prior to mapping. A wild card nucleotides "NNN"
(any of C, T, G, A) can for example be added for mapping non-perfect
reference hits. No nucleotides are added by default.}

\item{N_down}{Character. Same as N_up but indicating a sequence that should
be added to the reference at the 3' end. Useful for tRNA analysis where the
reference do not contain pre-processed tRNA. Setting N_down="NNN" or "CCA"
(in many species CCA is added to mature tRNA) will allow mapping against
the mature tRNA. No nucleotides are added by default.}

\item{report_string}{Logical whether an alignment string that shows in
character where sequences align against the reference. Works well with
tRNA, but makes the Alignments object difficult to work with when longer
references are used (default=FALSE).}
}
\value{
Stacked list, where each object on the highest level contains:
                   (Object 1) Reference name and sequence. 
                   (Object 2) Dataframe showing the mapping results of
                              each quiary sequence that mapped to Object 1.
}
\description{
\code{PAC_mapper} Mapping sequences against a reference.
}
\details{
Given a PAC object and the path to a fasta reference file, this function will
map sequences in PAC and mapping extract information.
}
\examples{


# More details on the examples can be found in the vignette.

library(seqpac)
load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", package = "seqpac", mustWork = TRUE))

###########################################################
### tRNA analysis in seqpac 
##----------------------------------------

# First create an annotation blanc PAC with group means
pac$Anno <- pac$Anno[,1, drop=FALSE]
pac_trna <- PAC_summary(pac, norm = "cpm", type = "means", pheno_target=list("stage"), merge_pac = TRUE)

# Then reannotate only tRNA using the PAC_mapper function
ref <- "/home/danis31/Desktop/Temp_docs/fasta/GtRNAdb/trna.fa"
map_object <- PAC_mapper(pac_trna, ref=ref, N_up = "NNN", N_down = "NNN", mapper="reanno", mismatches=0, threads=8, report_string=TRUE)


###########################################################
## Coverage plot of tRNA using PAC_covplot

# Single tRNA targeting a summary dataframe 
PAC_covplot(pac_trna, map=map_object, summary_target= list("cpmMeans_stage"), map_target="tRNA-Ala-AGC-1-1_chr3R:17657145-17657217_(+)")

# Find tRNAs with many fragments
n_tRFs <- unlist(lapply(map_object, function(x){nrow(x[[2]])}))
selct <- (names(map_object)[n_tRFs>1])[c(1, 16, 25, 43)]
cov_plt <- PAC_covplot(pac_trna, map=map_object, summary_target= list("cpmMeans_stage"), map_target=selct)
cowplot::plot_grid(plotlist=cov_plt, nrow=2, ncol=2)


###########################################################
## Analyze range types with map_rangetype and PAC_trna functions

# Download ss object from GtRNAdb 
dest_path <- file.path("/home/danis31/Desktop/Temp_docs/fasta/GtRNAdb/trna.tar.gz")
download.file(url="http://gtrnadb.ucsc.edu/genomes/eukaryota/Dmela6/dm6-tRNAs.tar.gz", destfile=dest_path)
untar(dest_path, exdir= dirname(dest_path), files = "dm6-tRNAs-confidence-set.ss")
ss_file <- "/home/danis31/Desktop/Temp_docs/fasta/GtRNAdb/dm6-tRNAs-confidence-set.ss"

# Classify fragments according to loop cleavage (small loops are omitted)       
map_object_ss <- map_rangetype(map_object, type="ss", ss=ss_file, min_loop_width=4)   # Generates warning         

# Remove reference tRNAs with no hits
map_object_ss <-  map_object_ss[!unlist(lapply(map_object_ss, function(x){x[[2]][1,1] == "no_hits"}))]
map_object_ss[[2]]


###########################################################
# Function classifying 5'-tRF, 5'halves, i-tRF, 3'-tRF, 3'halves

# Set tolerance for classification as a terminal tRF
tolerance <- 5  # 2 nucleotides from start or end of full-length tRNA)

# Apply the tRNA_class function and make a tRNA type column
pac_trna <- tRNA_class(pac_trna, map=map_object_ss, terminal=tolerance)
pac_trna$Anno$type <- paste0(pac_trna$Anno$decoder, pac_trna$Anno$acceptor)
head(pac_trna$Anno)

# Now use PAC_trna to generate some graphs based on grand means
trna_result <- PAC_trna(pac_trna, norm="cpm", filter = NULL,
  join = TRUE, top = 15, log2fc = TRUE,
  pheno_target = list("stage", c("Stage1", "Stage3")), 
  anno_target_1 = list("type"),
  anno_target_2 = list("class"))

cowplot::plot_grid(trna_result$plots$Expression_Anno_1$Grand_means,
                   trna_result$plots$Log2FC_Anno_1,
                   trna_result$plots$Percent_bars$Grand_means,
                   nrow=1, ncol=3)

# By setting join = FALSE you will get group means
trna_result <- PAC_trna(pac_trna, norm="cpm", filter = NULL,
  join = FALSE, top = 15, log2fc = TRUE,
  pheno_target = list("stage", c("Stage1", "Stage3")), 
  anno_target_1 = list("type"),
  anno_target_2 = list("class"))

cowplot::plot_grid(trna_result$plots$Expression_Anno_1$Stage1,
                   trna_result$plots$Expression_Anno_1$Stage3,
                   trna_result$plots$Log2FC_Anno_1,
                   trna_result$plots$Percent_bars$Stage1,
                   trna_result$plots$Percent_bars$Stage3,
                   nrow=1, ncol=5)
}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_gtf}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_pirna}()},
\code{\link{PAC_rpm}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_trna}()},
\code{\link{map_rangetype}()},
\code{\link{tRNA_class}()}
}
\concept{PAC analysis}
