% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_norm.R
\name{PAC_norm}
\alias{PAC_norm}
\title{Generates normalized values from a PAC object}
\usage{
PAC_norm(PAC, type = "rpm", PAC_merge = TRUE)
}
\arguments{
\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names and a Count table with raw counts.}

\item{type}{Character describing what type of normalization method to be
applied to the PAC$Counts. If type="rpm" will return reads per million (or
sometimes refered to as counts per million reads). Each sequence is then
divided against the total number of sequences in a given sample. If
type="vst" PAC$Counts will instead be imported into the
\code{varianceStabilizingTransformation} function in the DESeq2-package
with the options blind=TRUE and fitType="mean".
If type="rlog" PAC$Counts will instead be imported into the
\code{rlogTransformation} function in the DESeq2-package
with the options blind=TRUE and fitType="mean".}

\item{PAC_merge}{logical whether the normalized table should be returned as
stored in the PAC$norm folder of the provided PAC object (TRUE) or be
retruned as a data frame.}
}
\value{
A normalized count table
}
\description{
\code{PAC_norm} Generates RPM values from a PAC object
}
\details{
Using the counts in a PAC object to generate RPM values in a dataframe with
the same rownames as in the original PAC object
}
\examples{
  df  <- PAC_norm(PAC, type="rpm") 
}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_rpm}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_tRNA}()},
\code{\link{map_rangetype}()}
}
\concept{PAC analysis}
