% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_pca.R
\name{PAC_pca}
\alias{PAC_pca}
\title{Pie plot from PAC}
\usage{
PAC_pca(
  PAC,
  norm = "counts",
  type = "pheno",
  graphs = TRUE,
  pheno_target = NULL,
  anno_target = NULL,
  labels = NULL,
  ...
)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{norm}{Character indicating what type of data to be used. If
type="counts" the PCA will be conducted on the raw Counts. If type="rpm" the
analysis will be done on rpm values returned from \code{PAC_norm} function
and stored in the norm folder of the PAC-list object. The name of any other
table in the PAC$norm folder can also be used.}

\item{type}{Character indicating what type of pca and plot to be drawn. When
type="pheno" then results will be drawned from a sample perspective, while
if type="anno" then results will be presented from a sequence perspective.
If type="both" then a biplot will be drawn, representing both pheno and anno
features. Note, the 1st object (target column name) in the pheno_target or
anno_taget can be used to specifically highlight sample and sequence groups
(see below).}

\item{graphs}{Logical whether or not scatter plots should be plotted.
(default=TRUE)}

\item{pheno_target}{List with: 1st object being a character vector of target
column(s) in Pheno, 2nd object being a character vector of the target
group(s) in the target column (1st object). The 1st object also control
group colors when type="pheno" or "both". (default=NULL)}

\item{anno_target}{List with: 1st object being a character vector of target
column(s) in Anno, 2nd object being a character vector of the target
features(s) in the target column (1st object). The 1st object also control
sequence feature colors when type="anno". (default=NULL)}

\item{labels}{If labels="sample", then points will be labeled with the names
rownames in PAC$Pheno or PAC$Anno depending on type="pheno" or type="anno",
respectively. Point labels can also be manually provided as a character
vector in the same length as the intended target. As default, labels=NULL
where only point are plotted.}

\item{...}{parsing to the fviz_pca functions of the factoextra package.}
}
\value{
A PCA list object generated by the PCA function in the FactoMineR
 package
}
\description{
\code{PAC_pca} PAC principle component analysis.
}
\details{
Given a PAC object the function will perform a principle component analysis by
calling the PCA function in the FactoMineR package, and then plot scatter plots 
with the fviz_pca functions of the factoextra package.
}
\examples{

### Load data ###
library(seqpac)
load(system.file("extdata", "drosophila_sRNA_pac.Rdata", package = "seqpac", mustWork = TRUE))

# Add normalized data as counts per million reads (cpm): 
pac_master <- PAC_norm(pac_master, norm="cpm")

# Simple sample counts pca and scatterplots with no groupings: 
pca_cnt <- PAC_pca(pac_master, norm="counts")

# Sample cpm pca and scatterplots with color groupings from PAC$Pheno$type column:    
pca_cpm <- PAC_pca(pac_master, norm="cpm", type="pheno", pheno_target=list("stage"))

# Same but with or without text labels: 
pca_cpm_lab <- PAC_pca(pac_master, norm="cpm", type="pheno", pheno_target=list("stage"), labels="samples")
pca_cpm_lab2 <- PAC_pca(pac_master, norm="cpm", type="pheno", pheno_target=list("stage"), labels=pac_master$Pheno$batch)

# Cpm pca with PAC$Anno sequence features instead of Pheno samples and restricted to read size 20-22:
pca_cpm_anno <- PAC_pca(pac_master, norm="cpm", type="anno", anno_target=list("Size", 20:22))

# Cpm pca as biplot:
pca_cpm_bi <- PAC_pca(pac_master, norm="cpm", type="both", pheno_target=list("stage"))

# Plot individual graphs
pca_cpm_bi$graphs[[1]]
pca_cpm_lab$graphs[[1]]
pca_cpm_anno$graphs[[3]]

# Extract pca output
pca_cpm_anno$pca

}
\seealso{
\url{https://github.com/Danis102} for updates on the current package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_gtf}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_pirna}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_trna}()},
\code{\link{filtsep_bin}()},
\code{\link{map_rangetype}()},
\code{\link{tRNA_class}()}
}
\concept{PAC analysis}
