% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_pie.R
\name{PAC_pie}
\alias{PAC_pie}
\title{Pie plot from PAC}
\usage{
PAC_pie(
  PAC,
  anno_target = NULL,
  pheno_target = NULL,
  colors = NULL,
  no_anno = TRUE,
  summary = "sample",
  angle = -25
)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{anno_target}{List with: 
1st object being character vector of target column(s) in Anno, 
2nd object being a character vector of the target
biotype(s) in the target column (1st object).
Important, the 2nd object is order sensitive,
meaning that categories will appear in the same
order in the pie. (default=NULL)}

\item{pheno_target}{List with: 
1st object being character vector of target column(s) in Pheno, 
2nd object being a character vector of the target
group(s) in the target column (1st object).
Important, the 2nd object is order sensitive,
meaning that categories will appear in the same
order in the pie. (default=NULL)}

\item{colors}{Character vector RGB color codes as generated by for example
grDevices::colorRampPalette. If colors=NULL (default), colors will be
generated using a default color palett. Important: In default mode,
categories named "other" or no_anno" will automtically recieve a grey
color.}

\item{summary}{Character string defining whether to stack individual samples
or using a mean of samples. If summary="samples" individual samples will be
plotted (default). If summary="pheno" means of the sample groups targeted
by the pheno_target input will be plotted. If summary="all" mean of all
samples will be plotted.}

\item{angle}{Integer vector (poitive or negative) that sets the rotation of
the pie.}
}
\value{
A pie plot
}
\description{
\code{PAC_pie} analyses nucleotide bias.
}
\details{
Given a PAC object the function will summarize the counts into percent
biotype and plot a pie chart.
}
\examples{

##########################################
### Pie chart in seqpac 
##----------------------------------------


load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", package = "seqpac", mustWork = TRUE))

# Choose an anno_target and plot samples (summary="samples"; default)
output_pie <- PAC_pie(pac, anno_target=list("Biotypes_mis0"))
cowplot::plot_grid(plotlist=output_pie, scale = 0.8)

output_pie <- PAC_pie(pac, anno_target=list("Biotypes_mis3"))
cowplot::plot_grid(plotlist=output_pie, ncol=3, scale = 0.5)


# Make ordered pie charts of grand mean percent of all samples
ord_bio <- as.character(sort(unique(pac$Anno$Biotypes_mis3)), unique(pac$Anno$Biotypes_mis0))
output_pie_1 <- PAC_pie(pac, anno_target=list("Biotypes_mis0", ord_bio), summary="all")
output_pie_2 <- PAC_pie(pac, anno_target=list("Biotypes_mis3", ord_bio), summary="all")
cowplot::plot_grid(plotlist=c(output_pie_1, output_pie_2), nrow=2, scale = 1.2)

# Rotate
PAC_pie(pac, anno_target=list("Biotypes_mis0"), summary="all", angle=180)
PAC_pie(pac, anno_target=list("Biotypes_mis0"), summary="all", angle=40)

# Compare biotype mapping with or without mismaches and group by PAC$Pheno
ord_bio <- as.character(sort(unique(pac$Anno$Biotypes_mis0)))  # Make sure that both get same biotype order
output_mis0 <- PAC_pie(pac, pheno_target=list("stage"), summary="pheno", anno_target=list("Biotypes_mis0", ord_bio))
output_mis3 <- PAC_pie(pac, pheno_target=list("stage"), summary="pheno", anno_target=list("Biotypes_mis3", ord_bio))
cowplot::plot_grid(plotlist=c(output_mis0, output_mis3), labels = names(c(output_mis0, output_mis3)), nrow=2)
 
   
}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_gtf}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pirna}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_trna}()},
\code{\link{map_rangetype}()},
\code{\link{tRNA_class}()}
}
\concept{PAC analysis}
