% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_pie.R
\name{PAC_pie}
\alias{PAC_pie}
\title{Pie plot from PAC}
\usage{
PAC_pie(PAC, pheno_target, anno_target, colvec = NULL, angle = -25)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{pheno_target}{List with: 
1st object being character vector of target column(s) in Pheno, 
2nd object being a character vector of the target
group(s) in the target column (1st object).
Important, the 2nd object is order sensitive,
meaning that categories will appear in the same
order in the pie. (default=NULL)}

\item{anno_target}{List with: 
1st object being character vector of target column(s) in Anno, 
2nd object being a character vector of the target
biotype(s) in the target column (1st object).
Important, the 2nd object is order sensitive,
meaning that categories will appear in the same
order in the pie. (default=NULL)}

\item{colvec}{Character vector RGB color codes as generated by for example
grDevices::colorRampPalette. If colvec=NULL (default), colors will be
generated using a default color palett. Important: In default mode,
categories named "other" or no_anno" will automtically recieve a grey
color.}

\item{angle}{Integer vector (poitive or negative) that sets the rotation of
the pie.}
}
\value{
A pie plot
}
\description{
\code{PAC_pie} analyses nucleotide bias.
}
\details{
Given a PAC object the function will summarize the counts into percent
biotype and plot a pie chart.
}
\examples{

### Load data ###
path="/data/Data_analysis/Projects/Drosophila/Other/IOR/Jan_IOR_200130/R_analysis_full/"
load(file=paste0(path, "PAC_all.Rdata"))

### Generate rpm values ###
PAC_all <- PAC_rpm(PAC_all)

### Filter PAC file ###
PAC_filt <- PAC_filter(PAC_all, size = c(16, 45), threshold = 10, coverage = 50, type = "rpm", stat = TRUE, pheno_target = list("Method", c("IOR1_proto", "IOR1_tgirt")), anno_target = NULL)
PAC_filt <- PAC_filter(PAC_all, size = c(16, 45), threshold = 10, coverage = 10, type = "rpm", stat = TRUE, pheno_target = NULL, anno_target = NULL)

### Generate hierachy of simplified biotypes ###
unique(do.call("c", strsplit(as.character(PAC_filt$Anno$mis0), ";|\\\\|")))

hierachy = list( rRNA="rRNA",
                 Mt_tRNA= "tRNA_mt-tRNA",
                 tRNA="Ensembl_tRNA|tRNA_nuc-tRNA",
                 miRNA="^miRNA|Ensembl_miRNA",
                 pre_miRNA="Ensembl_pre_miRNA",
                 snoRNA="snoRNA",
                 snRNA="snRNA",
                 lincRNA="lincRNA",
                 piRNA="piRNA")
                 
hierachy = list( rRNA="rRNA",
                 tRNA="Ensembl_tRNA|tRNA_nuc-tRNA|tRNA_mt-tRNA",
                 miRNA="^miRNA|Ensembl_miRNA|Ensembl_pre_miRNA",
                 piRNA="piRNA")
                 
PAC_filt <- simplify_reanno(PAC_filt, hierachy=hierachy, mismatches=0, bio_name="Biotypes", PAC_merge=TRUE)

### Plot stacked bargraphs
table(PAC_filt$Anno$Biotypes)
unique(PAC_filt$Anno$Biotypes)
biotypes_ord <- c("rRNA", "tRNA", "Mt_tRNA", "miRNA", "piRNA", "snoRNA", "snRNA", "lincRNA", "no_anno")
biotypes_ord <- c("rRNA", "tRNA", "miRNA", "piRNA", "other", "no_anno")

PAC_filt$Pheno$Method_Tag <- gsub("__", "", paste(PAC_filt$Pheno$Method, PAC_filt$Pheno$Tag, sep="_"))
table(PAC_filt$Pheno$Method_Tag)
unique(PAC_filt$Pheno$Method_Tag)
method_ord <- c("R2_R2R","IOR1_proto","IOR1_tgirt","IOR2_Long_proto_Tag","IOR2_Short_proto_Tag", "IOR2_Long_proto_No_Tag", "IOR2_Short_proto_No_Tag")

plot_lst  <- PAC_pie(PAC_filt, anno_target = list("Biotypes", biotypes_ord), pheno_target = list("Method_Tag", method_ord), colvec=NULL, angle=-25)  


cowplot::plot_grid(plotlist=plot_lst, labels=names(plot_lst))
 
   
}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: \code{\link{PAC_covplot}},
  \code{\link{PAC_deseq}}, \code{\link{PAC_filter}},
  \code{\link{PAC_jitter}}, \code{\link{PAC_mapper}},
  \code{\link{PAC_nbias}}, \code{\link{PAC_pca}},
  \code{\link{PAC_rpm}}, \code{\link{PAC_saturation}},
  \code{\link{PAC_stackbar}}, \code{\link{PAC_summary}}
}
\concept{PAC analysis}
