% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_rpm.R
\name{PAC_rpm}
\alias{PAC_rpm}
\title{Generates normalized values from a PAC object (deprecated)}
\usage{
PAC_rpm(PAC)
}
\arguments{
\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names and a Count table with raw counts.}
}
\value{
A normalized count table (
}
\description{
\code{PAC_rpm} Generates RPM values from a PAC object (deprecated)
}
\details{
Using the counts in a PAC object to generate RPM values in a dataframe with
the same rownames as in the original PAC object
}
\examples{
  df  <- PAC_norm(PAC, type="rpm") 
}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_tRNA}()},
\code{\link{map_rangetype}()}
}
\concept{PAC analysis}
