% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_saturation.R
\name{PAC_saturation}
\alias{PAC_saturation}
\title{Filter a PAC object on sequence size and covarage}
\usage{
PAC_saturation(PAC, resample = 10, steps = 10, thresh = c(1, 500),
  cumulative = FALSE, start_perc = 1, threads = 8)
}
\arguments{
\item{PAC}{PAC-list object containing a Counts data.frame with sequences as
row names and samples as column names.}

\item{resample}{Integer setting the number of permutations at each percentage
step (default=10).}

\item{steps}{Integer the number of percentage steps between 0-100% of the
original dataset (default=10).}

\item{thresh}{Integer vector containing two mean count thresholds that will
be targeted. Default is set to c(1,500), where each new occurance (>=1) and
each new occurance reaching 500 counts (>=500) will be analyzed.}

\item{cumulative}{(optional) Logical. FALSE make the analysis on total unique
sequences when resampling at different percentages (default). TRUE make the
analysis on unique sequences gained at each step compared to a given
downsampled dataset given in 'start_perc' (default = 1%). While the integer
in the 'resample' input sets how many (n) permutations to be made at each
step, permutation with FALSE will permute n times at each percentage step,
while TRUE will permute 1 time per step and then repeat this n times. The
two approaches gives similar results.}

\item{start_perc}{(optional) Logical specifying if an coverage graph should
be generated or not (default=FALSE).}

\item{threads}{Number of cores to be used for performing the permutations.}
}
\value{
A list with two graph objects: The 1:st graph (A) shows
  saturation/diversity result at the 1:st threshold. The 2:nd graph (B) shows
  saturation/diversity result at the 2:nd threshold.
}
\description{
\code{PAC_saturation} Perfoms an sequence diversity/saturation analysis on a
PAC objects.
}
\details{
Given a PAC object the function will perform a sequence saturation analysis.
This is done by downsampling the original dataset by permutation at different
percentages of the original dataset. The flatter the curve at the original
sequence depth (100%) the more saturated is the diversity of sequences at the
original dataset.
}
\examples{

library(seqpac)
path="/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/"
load(file=paste0(path, "PAC_all.Rdata"))

plot_lst  <- PAC_saturation(PAC, resample=10, steps=10, thresh=c(1,500), cumulative=TRUE, start_perc=1, threads=8)
cowplot::plot_grid(plot_lst$A,plot_lst$B)

}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: \code{\link{PAC_covplot}},
  \code{\link{PAC_deseq}}, \code{\link{PAC_filter}},
  \code{\link{PAC_jitter}}, \code{\link{PAC_mapper}},
  \code{\link{PAC_mismap}}, \code{\link{PAC_nbias}},
  \code{\link{PAC_pca}}, \code{\link{PAC_pie}},
  \code{\link{PAC_rpm}}, \code{\link{PAC_sizedist}},
  \code{\link{PAC_stackbar}}, \code{\link{PAC_summary}}
}
\concept{PAC analysis}
