% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_sizedist.R
\name{PAC_sizedist}
\alias{PAC_sizedist}
\title{Generates a size distribution analysis from sequences and counts in a PAC object}
\usage{
PAC_sizedist(PAC, range = NULL, anno_target, pheno_target = NULL,
  colvect = NULL)
}
\arguments{
\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names and a count table with raw counts.}

\item{range}{Integer vector giving the range  in sequence lengths (default=c(min, max)).#'}

\item{anno_target}{List with: 
1st object being character vector of target column(s) in Anno, 
2nd object being a character vector of the target biotypes(s) in the target column (1st object).
(default=NULL)}

\item{pheno_target}{List with: 
1st object being character vector of target column(s) in Pheno, 
2nd object being a character vector of the target group(s) in the target column (1st object).
(default=NULL)}
}
\value{
A list of objects: 
              1st object (Histograms::Samples): Individual histograms showing the nucleotide ratios per sample over the specified range.   
              2nd object (Data::Samples): Data used to generate the plots.
              3rd object (Stacked_bars::Groups): Stacked bars showing the mean ratios of each nucleotide per group over the specified range.
              4th object (Error_bars::Groups): Error bar plots with mean ratio of each nucleotide per group over the specified range.
}
\description{
\code{PAC_sizedist} analyses nucleotide bias.
}
\details{
Given a PAC object the function will attempt to extract the ratios of
specific biotypes at a given position in sequences of the Anno data.frame
in relation to the sequence counts in Counts.
}
\examples{

library(seqpac)
path="/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/"
load(file=paste0(path, "PAC_all.Rdata"))

### First apply shallow counts filter, and then generate rpm on which rpm filter is applied.  
PAC_filt <- PAC_filter(PAC_all, size=c(16,45), threshold=10, coverage=5, type="counts", stat=FALSE, pheno_target=NULL, anno_target=NULL)
PAC_filt <- PAC_rpm(PAC_filt)
PAC_filt <- PAC_filter(PAC_filt, size=c(16,45), threshold=10, coverage=5, type="rpm", stat=FALSE, pheno_target=NULL, anno_target=NULL)

hierachy <- list( Mt_rRNA= "12S|16S|Mt_rRNA",
                 rRNA="5S|5.8S|18S|28S|S45|Ensembl_rRNA|rRNA_Other",
                 Mt_tRNA= "tRNA_mt-tRNA",
                 tRNA="Ensembl_tRNA|tRNA_nuc-tRNA",
                 miRNA="^miRNA|Ensembl_miRNA|Ensembl_pre_miRNA",
                 piRNA="piRNA")
                 
hierachy <- list(miRNA="^miRNA|Ensembl_miRNA|Ensembl_pre_miRNA",
                 Mt_rRNA= "12S|16S|Mt_rRNA",
                 rRNA="5S|5.8S|18S|28S|S45|Ensembl_rRNA|rRNA_Other",
                 Mt_tRNA= "tRNA_mt-tRNA",
                 tRNA="Ensembl_tRNA|tRNA_nuc-tRNA",
                 piRNA="piRNA")
                 

PAC_filt <- simplify_reanno(PAC_filt, hierachy=hierachy, mismatches=0, bio_name="Biotypes_1", PAC_merge = TRUE)

PAC_filt$Pheno$Groups <- paste(PAC_filt$Pheno$Batch, PAC_filt$Pheno$Method, PAC_filt$Pheno$Method_tag, PAC_filt$Pheno$Tag, sep="_")


## Plot only histograms
sizedist_result <- PAC_sizedist(PAC_filt, anno_target=list("Biotypes_1", c("no_anno", "other", "miRNA", "tRNA", "Mt_tRNA", "rRNA", "Mt_rRNA",  "piRNA")), 
pheno_target=list("Groups", unique(PAC_filt$Pheno$Groups)))

sizedist_result <- PAC_sizedist(PAC=PAC_filt, anno_target=list("Biotypes_1"))
      
filt <- PAC_filt$Pheno$Groups \%in\% c("Sep_Proto_POOH__","Sep_Proto_Long_POOH_tag","Sep_Proto_Short_POOH_tag") 
filt <- PAC_filt$Pheno$Groups \%in\% c("Sep_Proto_POOH__","Jan_Proto_Long_POOH_tag","Jan_Proto_Short_POOH_tag")
filt <- PAC_filt$Pheno$Groups \%in\% c("Sep_Proto_IOR__","Jan_Proto_IOR__","Sep_TGIRT_POOH__")
filt <- PAC_filt$Pheno$Groups \%in\% c("Sep_Proto_Long_IOR_notag","Jan_Proto_Long_IOR_notag","Sep_Proto_Short_IOR_notag") 
filt <- PAC_filt$Pheno$Groups \%in\% c("Sep_Proto_Short_IOR_notag","Jan_TGIRT_IOR__","Sep_TGIRT_IOR__") 
filt <- PAC_filt$Pheno$Groups \%in\% c("Sep_Proto_Short_IOR_notag","Sep_TGIRT_POOH__","Jan_R2_R2R_TGIRT_R2_R2R__") 

cowplot::plot_grid(plotlist=sizedist_result[[1]][filt], nrow = 3, ncol = 3)

cowplot::plot_grid(plotlist=sizedist_result[[1]][37:45], nrow = 3, ncol = 3)



}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: \code{\link{PAC_covplot}},
  \code{\link{PAC_deseq}}, \code{\link{PAC_filter}},
  \code{\link{PAC_jitter}}, \code{\link{PAC_mapper}},
  \code{\link{PAC_mismap}}, \code{\link{PAC_nbias}},
  \code{\link{PAC_pca}}, \code{\link{PAC_pie}},
  \code{\link{PAC_rpm}}, \code{\link{PAC_saturation}},
  \code{\link{PAC_stackbar}}, \code{\link{PAC_summary}},
  \code{\link{PAC_trna}}
}
\concept{PAC analysis}
