% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_sizedist.R
\name{PAC_sizedist}
\alias{PAC_sizedist}
\title{Generates size distribution plots from sequences and counts in a PAC object}
\usage{
PAC_sizedist(
  PAC,
  norm = "counts",
  range = NULL,
  anno_target,
  pheno_target = NULL,
  summary_target = NULL,
  colors = NULL
)
}
\arguments{
\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names and a count table with raw counts.}

\item{range}{Integer vector giving the range  in sequence lengths (default=c(min, max)).#'}

\item{anno_target}{List with: 
1st object being character vector of target column(s) in Anno, 
2nd object being a character vector of the target biotypes(s) in the target column (1st object).
(default=NULL)}

\item{pheno_target}{List with: 
1st object being character vector of target column(s) in Pheno, 
2nd object being a character vector of the target group(s) in the target column (1st object).
(default=NULL)}

\item{summary_target}{List with: 
1st object being character target object in PAC$summary, 
2nd object being a character vector of the target columns(s) in the target object (1st object).
(default=NULL)}
}
\value{
A list of objects: 
              1st object (Histograms::Samples): Individual histograms showing the nucleotide ratios per sample over the specified range.   
              2nd object (Data::Samples): Data used to generate the plots.
              3rd object (Stacked_bars::Groups): Stacked bars showing the mean ratios of each nucleotide per group over the specified range.
              4th object (Error_bars::Groups): Error bar plots with mean ratio of each nucleotide per group over the specified range.
}
\description{
\code{PAC_sizedist} plotting nucleotide bias.
}
\details{
Given a PAC object the function will attempt to order sequences by their size
(number of nucleotides) and extract the ratios of specific classifications
(e.g. sRNA classes) at each size point.
}
\examples{

##########################################
### Stacked bars in seqpac 
##----------------------------------------
library(seqpac)
load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", package = "seqpac", mustWork = TRUE))


# Choose an anno_target and plot samples (summary="samples")
PAC_stackbar(pac, anno_target=list("Biotypes_mis0"))

# 'no_anno' and 'other' will always end on top not matter the order
ord_bio <- as.character(sort(unique(pac$Anno$Biotypes_mis3)))
p1 <- PAC_stackbar(pac, anno_target=list("Biotypes_mis0", ord_bio))
p2 <- PAC_stackbar(pac, anno_target=list("Biotypes_mis0", rev(ord_bio)))
cowplot::plot_grid(plotlist=list(p1, p2))

# Reorder samples by pheno_targets
PAC_stackbar(pac, pheno_target=list("batch"), summary="samples", anno_target=list("Biotypes_mis0"))

# Summarized over pheno_target 
# (as default PAC_stackbar orders by pheno_target but plots all samples, unless summary="pheno")
PAC_stackbar(pac, anno_target=list("Biotypes_mis0"), summary="pheno", pheno_target=list("stage"))


}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_gtf}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_pirna}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_trna}()},
\code{\link{map_rangetype}()},
\code{\link{tRNA_class}()}
}
\concept{PAC analysis}
