% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_sizedist.R
\name{PAC_sizedist}
\alias{PAC_sizedist}
\title{Generates a size distribution analysis from sequences and counts in a PAC object}
\usage{
PAC_sizedist(PAC, norm = NULL, range = NULL, anno_target,
  pheno_target = NULL, summary_target = NULL, colors = NULL)
}
\arguments{
\item{PAC}{PAC-list object containing an Anno data.frame with sequences as
row names and a count table with raw counts.}

\item{range}{Integer vector giving the range  in sequence lengths (default=c(min, max)).#'}

\item{anno_target}{List with: 
1st object being character vector of target column(s) in Anno, 
2nd object being a character vector of the target biotypes(s) in the target column (1st object).
(default=NULL)}

\item{pheno_target}{List with: 
1st object being character vector of target column(s) in Pheno, 
2nd object being a character vector of the target group(s) in the target column (1st object).
(default=NULL)}

\item{summary_target}{List with: 
1st object being character target object in PAC$summary, 
2nd object being a character vector of the target columns(s) in the target object (1st object).
(default=NULL)}
}
\value{
A list of objects: 
              1st object (Histograms::Samples): Individual histograms showing the nucleotide ratios per sample over the specified range.   
              2nd object (Data::Samples): Data used to generate the plots.
              3rd object (Stacked_bars::Groups): Stacked bars showing the mean ratios of each nucleotide per group over the specified range.
              4th object (Error_bars::Groups): Error bar plots with mean ratio of each nucleotide per group over the specified range.
}
\description{
\code{PAC_sizedist} analyses nucleotide bias.
}
\details{
Given a PAC object the function will attempt to extract the ratios of
specific biotypes at a given position in sequences of the Anno data.frame
in relation to the sequence counts in Counts.
}
\examples{


library(seqpac)
load("/home/danis31/OneDrive/Programmering/Programmering/Pipelines/Drosophila/Pipeline_3.1/seqpac/dm_test_PAC.Rdata")

PAC_filt <- PAC_rpm(PAC_filt)
PAC_filt <- PAC_summary(PAC=PAC_filt, norm = "rpm", type = "means", pheno_target=list("Method"))

hierarchy <- list( Mt_rRNA= "12S|16S|Mt_rRNA",
                rRNA="5S|5.8S|18S|28S|S45|Ensembl_rRNA|rRNA_Other",
                Mt_tRNA= "tRNA_mt-tRNA",
               tRNA="Ensembl_tRNA|tRNA_nuc-tRNA",
               miRNA="^miRNA|Ensembl_miRNA|Ensembl_pre_miRNA",
               piRNA="piRNA")
hierarchy <- hierarchy[c(5,1,2,3,4,6)]          
as.data.frame(names(hierarchy))              

PAC_filt <- simplify_reanno(PAC_filt, hierarchy=hierarchy, mismatches=0, bio_name="Biotypes_mis0", PAC_merge=TRUE)

ord <- c("no_anno", "other", "miRNA", "tRNA", "Mt_tRNA", "rRNA", "Mt_rRNA",  "piRNA")

sizedist_plots <- PAC_sizedist(PAC_filt, anno_target=list("Biotypes_mis0", ord), summary_target=list("rpmMeans_Method"))
cowplot::plot_grid(plotlist=sizedist_plots[[1]], nrow = 2, ncol = 1)

sizedist_plots <- PAC_sizedist(PAC_filt, norm="raw", anno_target=list("Biotypes_mis0", ord), pheno_target=list("Method", "TGIRT"))
cowplot::plot_grid(plotlist=sizedist_plots[[1]], nrow = 2, ncol = 2)


}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: \code{\link{PAC_covplot}},
  \code{\link{PAC_deseq}}, \code{\link{PAC_filter}},
  \code{\link{PAC_filtsep}}, \code{\link{PAC_jitter}},
  \code{\link{PAC_mapper}}, \code{\link{PAC_mismap}},
  \code{\link{PAC_nbias}}, \code{\link{PAC_pca}},
  \code{\link{PAC_pie}}, \code{\link{PAC_rpm}},
  \code{\link{PAC_saturation}}, \code{\link{PAC_stackbar}},
  \code{\link{PAC_summary}}, \code{\link{PAC_trna}}
}
\concept{PAC analysis}
