% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_stackbar.R
\name{PAC_stackbar}
\alias{PAC_stackbar}
\title{Plot stackbars from PAC}
\usage{
PAC_stackbar(
  PAC,
  anno_target = NULL,
  pheno_target = NULL,
  color = NULL,
  width = 1,
  no_anno = TRUE,
  total = TRUE,
  summary = "samples"
)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{anno_target}{List with: 
1st object being character vector of target
column(s) in Anno, 2nd object being a character
vector of the target biotype(s) in the target column
(1st object). Important, the 2nd object is order
sensitive, meaning that categories will appear in
the same order in the stacked bargraph.
(default=NULL)}

\item{pheno_target}{List with: 
1st object being character vector of target
column(s) in Pheno, 2nd object being a character
vector of the target group(s) in the target column
(1st object). Important, the 2nd object is order
sensitive, meaning that categories will appear in
the same order in the stacked bargraph.
(default=NULL)}

\item{color}{Character vector with rgb colour hex codes in the same length
as the number of biotypes. For example see:
https://www.coolgenerator.com/rgb-color-generator. color=NULL will
generate the default color scheme.}

\item{width}{Integer adjusting the width of the bars (default=1)}

\item{no_anno}{Logical. If TRUE sequences with no annotations will be
plotted, while FALSE will skip sequences with 'no_anno' in the column
defined by anno_target (default=TRUE). Note that you can always use
\code{PAC_filter} to remove anno_targets from PAC prior to plotting.}

\item{total}{Logical, whether the total counts should be added at the bottom
of each graph (default=TRUE).}

\item{summary}{Character vector defining whether to stack individual samples
as one stack, or using a group mean of samples sharing the same value in
the specified pheno_target. Accepts, "samples" (individual samples), "pheno"
(means of pheno_target) or "all" (mean of al samples) as input.
(Default="samples").}
}
\value{
A stacked bar plot generated by ggplot2
}
\description{
\code{PAC_stackbar} analyses nucleotide bias.
}
\details{
Given a PAC object the function will attempt to extract group information
from Pheno, biotype information from Anno, and summaries this over the data
in Counts or norm.
}
\examples{

library(seqpac)
load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                 package = "seqpac", mustWork = TRUE))

##########################################
### Stacked bars in seqpac 
##----------------------------------------

# Choose an anno_target and plot samples (summary="samples")
PAC_stackbar(pac, anno_target=list("Biotypes_mis0"))

# 'no_anno' and 'other' will always end on top not matter the order
ord_bio <- as.character(sort(unique(pac$Anno$Biotypes_mis3)))
p1 <- PAC_stackbar(pac, anno_target=list("Biotypes_mis0", ord_bio))
p2 <- PAC_stackbar(pac, anno_target=list("Biotypes_mis0", rev(ord_bio)))
cowplot::plot_grid(plotlist=list(p1, p2))
# (Hint: if you want them to appear not on top, rename them)

# Reorder samples by pheno_targets
PAC_stackbar(pac, pheno_target=list("batch"), summary="samples", 
             anno_target=list("Biotypes_mis0"))

# Summarized over pheno_target 
# (as default PAC_stackbar orders by pheno_target but plots all samples, 
#  unless summary="pheno")
PAC_stackbar(pac, anno_target=list("Biotypes_mis0"), 
             summary="pheno", pheno_target=list("stage"))

# Summerized over a grand mean of all samples
PAC_stackbar(pac, anno_target=list("Biotypes_mis0"), summary="all")
}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_gtf}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_trna}()},
\code{\link{filtsep_bin}()},
\code{\link{map_rangetype}()},
\code{\link{tRNA_class}()}
}
\concept{PAC analysis}
