% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_stackbar.R
\name{PAC_stackbar}
\alias{PAC_stackbar}
\title{Plot stackbars from PAC}
\usage{
PAC_stackbar(
  PAC,
  anno_target = list("Biotypes", NULL),
  pheno_target = NULL,
  colvec = NULL,
  width = 1,
  no_anno = TRUE,
  total = FALSE
)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{anno_target}{List with: 
1st object being character vector of target column(s) in Anno, 
2nd object being a character vector of the target
biotype(s) in the target column (1st object).
Important, the 2nd object is order sensitive,
meaning that categories will appear in the same
order in the stacked bargraph. (default=NULL)}

\item{pheno_target}{List with: 
1st object being character vector of target column(s) in Pheno, 
2nd object being a character vector of the target
group(s) in the target column (1st object).
Important, the 2nd object is order sensitive,
meaning that categories will appear in the same
order in the stacked bargraph. (default=NULL)}

\item{colvec}{Character vector with rgb colour hex codes in the same length
as the number of biotypes. For example see:
https://www.coolgenerator.com/rgb-color-generator. colvec=NULL will
generate the default color scheme.}

\item{width}{Integer adjusting the width of the bars (default=1)}

\item{no_anno}{Logical. If TRUE sequences with no annotations will be
plotted, while FALSE will skip sequences with no annotation in the column
defined by anno_target (default=TRUE).}

\item{total}{Logical, whethter the total counts should be added to each bar
or not (default=FALSE).}
}
\value{
A stacked bar plot generated by ggplot2
}
\description{
\code{PAC_stackbar} analyses nucleotide bias.
}
\details{
Given a PAC object the function will attempt to extract group information
from Pheno, biotype information from Anno, and summaries this over the data
in Counts or norm.
}
\examples{

load(file="/data/Data_analysis/Projects/Drosophila/Other/IOR/Joint_analysis/R_analysis/PAC_hifilt.Rdata")


#-------------------------------------------------------------------------------------------------#
#  Stacked biotype barplots 
#-------------------------------------------------------------------------------------------------#

biotypes_ord <- c("rRNA","Mt_rRNA", "tRNA", "Mt_tRNA", "miRNA", "piRNA","other", "no_anno")

method_ord <- unique(PAC_hifilt$Pheno$Groups)[c(1, 15, 3 ,7, 9, 6, 14, 12, 8, 5, 13, 11, 2, 10, 4)]

plot_mis0  <- PAC_stackbar(PAC_hifilt, anno_target = list("Biotypes_mis0", biotypes_ord), pheno_target = list("Groups", method_ord), width=1, no_anno=TRUE, total=TRUE)
plot_mis3  <- PAC_stackbar(PAC_hifilt, anno_target = list("Biotypes_mis3", biotypes_ord), pheno_target = list("Groups", method_ord), width=1, no_anno=TRUE, total=TRUE)

table(PAC_IOR1$Anno$Biotypes_simpl)


}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_rpm}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_tRNA}()},
\code{\link{map_rangetype}()}
}
\concept{PAC analysis}
