% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_stackbar.R
\name{PAC_stackbar}
\alias{PAC_stackbar}
\title{Plot stackbars from PAC}
\usage{
PAC_stackbar(
  PAC,
  anno_target = NULL,
  pheno_target = NULL,
  color = NULL,
  width = 1,
  no_anno = TRUE,
  total = TRUE,
  summary = "samples"
)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{anno_target}{List with: 
1st object being character vector of target column(s) in Anno, 
2nd object being a character vector of the target
biotype(s) in the target column (1st object).
Important, the 2nd object is order sensitive,
meaning that categories will appear in the same
order in the stacked bargraph. (default=NULL)}

\item{pheno_target}{List with: 
1st object being character vector of target column(s) in Pheno, 
2nd object being a character vector of the target
group(s) in the target column (1st object).
Important, the 2nd object is order sensitive,
meaning that categories will appear in the same
order in the stacked bargraph. (default=NULL)}

\item{color}{Character vector with rgb colour hex codes in the same length
as the number of biotypes. For example see:
https://www.coolgenerator.com/rgb-color-generator. color=NULL will
generate the default color scheme.}

\item{width}{Integer adjusting the width of the bars (default=1)}

\item{no_anno}{Logical. If TRUE sequences with no annotations will be
plotted, while FALSE will skip sequences with 'no_anno' in the column
defined by anno_target (default=TRUE). Note that you can always use
\code{PAC_filter} to remove anno_targets from PAC prior to plotting.}

\item{total}{Logical, whether the total counts should be added at the bottom
of each graph (default=TRUE).}

\item{summary}{Character vector defining whether to stack individual 
samples as one stack, or a group of samples sharing the same value in
a specified pheno_target. Accepts "samples" or "pheno" as input. 
(Default="samples").}
}
\value{
A stacked bar plot generated by ggplot2
}
\description{
\code{PAC_stackbar} analyses nucleotide bias.
}
\details{
Given a PAC object the function will attempt to extract group information
from Pheno, biotype information from Anno, and summaries this over the data
in Counts or norm.
}
\examples{

library(seqpac)
load(system.file("extdata", "drosophila_sRNA_pac_anno.Rdata", package = "seqpac", mustWork = TRUE))

head(pac$Anno)
lapply(pac, tibble::as_tibble)


#-------------------------------------------------------------------------------------------------#
#  Stacked biotype barplots 
#-------------------------------------------------------------------------------------------------#

levels(pac$Anno$Biotypes_mis0)
biotypes_ord <- c("rRNA","Mt_rRNA", "tRNA", "Mt_tRNA", "miRNA", "snoRNA", "lncRNA","other", "no_anno") # Change biotype order
group_ord <- as.character(unique(pac$Pheno$type)[c(2,1)]) # Change group order

plot_mis0  <- PAC_stackbar(pac, anno_target = list("Biotypes_mis0", biotypes_ord), pheno_target = list("type", group_ord), width=1, no_anno=TRUE, total=TRUE)
plot_mis3  <- PAC_stackbar(pac, anno_target = list("Biotypes_mis3", biotypes_ord), pheno_target = list("type", group_ord), width=1, no_anno=TRUE, total=TRUE)


}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_gtf}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_pirna}()},
\code{\link{PAC_rpm}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_trna}()},
\code{\link{map_rangetype}()},
\code{\link{tRNA_class}()}
}
\concept{PAC analysis}
