% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC_trna.R
\name{PAC_tRNA}
\alias{PAC_tRNA}
\title{Pairwise tsRNA analysis}
\usage{
PAC_tRNA(
  PAC,
  norm = "rpm",
  filter = 100,
  join = FALSE,
  top = 15,
  log2fc = FALSE,
  pheno_target = NULL,
  anno_target_1 = NULL,
  ymax_1 = NULL,
  anno_target_2 = NULL,
  paired = FALSE,
  paired_IDs = NULL
)
}
\arguments{
\item{PAC}{PAC-list object. The Anno object needs to contain at least two
tsRNA classification columns, typically isotype (e.g. LysCTT, GlyGCC) and
rangetype (5'tsRNA, 5'tsRNA-halves, i'-tsRNA,  3'tsRNA, 3'tsRNA-halves).
These can be generated by first mapping sequences to tRNA reference using
\emph{PAC_mapping} and extract isotypes from tRNA the names, and then
obtain rangetype classifications using \emph{map_rangetype}.}

\item{norm}{Character indicating what type of data in PAC to be used as
input. If norm="raw" the raw counts in Counts will be used. Given any other
character string, the function will search for the string as a name on a
dataframe stored in the PAC$norm list-folder (created for example by PAC_rpm),
(default="rpm")}

\item{filter}{Integer specifying the minimum coverage value (specified by
norm) for a sequence to be included in the analysis. If filter=10 and
norm="rpm", only sequences that have at least 10 rpm in 100% of samples
will be included. (default=NULL).}

\item{join}{Logical, whether seperate expression tables of anno_target_1 and
anno_target_2 should be generated for each group in pheno_target (FALSE),
or if pheno_target groups should be joined (TRUE).}

\item{top}{Integer, specifying the number of the most highly expressed
classifications in anno_target_1 that should be reported in the graphs.}

\item{log2fc}{Logical, whether log2 fold change point-bars (independent
pheno_target groups) or errorbars (paired pheno_target groups) should be
reported.}

\item{pheno_target}{List with: 1st object being a character vector of the
target column name in Pheno, and the 2nd object being a character vector
specifying the names of two target groups in the target column (1st
object). Note, that this function will only work using two groups from the
pheno_target column, like this: pheno_target=list(<Pheno_column_name>,
c(<group_1>, <group_2>)). If there are more groups than two in the target
column, unspecified groups will not be analyzed.}

\item{anno_target_1}{List with: 1st object being a character of the target
column name in Anno, 2nd object being a character vector of the target
classifications in the 1st target Anno column. The anno_target_1 is
typically used for isotype classification in tsRNA analysis (example:
LysCTT, GlyGGC etc). Specifying only the column name, will automatically
include all classifications in the anno_target column.}

\item{anno_target_2}{Same as anno_target_2 but specifying the the second
classification typically used for 3'-5' classification in tsRNA analysis
(example: 5'tsRNA, 5'tsRNA-halves, i'-tsRNA,  3'tsRNA, 3'tsRNA-halves).
Note, that anno_target_2[[2]] is order sensitive, meaning that the order
will be perserved in the graphs.}

\item{paired}{Logical, whether pheno_target is given as paired samples (e.g.
before and after treatment of the same patient). Only works with
\\code{paired_IDs}.}

\item{paired_IDs}{Character, specifying the name of a column in Pheno
containing the paried sample names. For example, if pheno_target=list("treatment",
c("before", "after")), and paired=TRUE, then paired_IDs="patient", must be
a column in Pheno containing the same patient ID reported twice for each
patient (before and after).}
}
\value{
List of ggplot2 plots and the data used for generating the plots. Use
  ls.str() to explore each level.
}
\description{
\code{PAC_tRNA} Analyzing and plotting tsRNAs between two groups.
}
\details{
Given a PAC and a PAC_map object generated by \emph{PAC_mapping} this
function will attempt to analyze and plot tsRNA over two pheno_target groups,
using two tsRNA classifications (anno_target_1, annot_target_2).
}
\examples{



}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_rpm}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{map_rangetype}()}
}
\concept{PAC analysis}
