% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{pheno}
\alias{pheno}
\alias{pheno<-}
\alias{pheno,PAC-method}
\title{sample information}
\usage{
pheno(object)

pheno(object) <- value

\S4method{pheno}{PAC}(object)
}
\description{
sample information
}
