% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_report.R
\name{progress_report}
\alias{progress_report}
\title{Genereates a Progress Report from make_counts logs}
\usage{
progress_report(input, type = "seqpac")
}
\arguments{
\item{input}{Character containing Sports output path or an counts list object
generated by \code{make_counts}.}

\item{type}{Character indicating what type of input. If type="seqpac", input
should be a counts list object generated by \code{make_counts} with
cutadapt=TRUE, which contains a $counts table and $logs list with logs
generated by cutadapt and fastq_quality_filter. Importantly, it is not
possible to generate progress report from make_counts if cutadapt=FALSE.
If type="sports", input needs to be a path to a folder with Sports output
files. The function will search this path for the the progress_report
folder that is automatically generated by Sports running with the -k option
(save all intermediate files).}
}
\value{
data.frame
}
\description{
\code{progress_report} generates summarized progress report
from make_counts logs or intermediate files generated by Sports.
}
\details{
Given a list object generated by \code{make_counts} or the path to a Sports
output directory, this function will store overview statistics in a single
dataframe. This dataframe can easily be included in a Pheno table of a PAC
object by using \code{make_pheno}.
}
\section{Important}{
 Progress reports will only be generated if using the -k
  option in Sports or cutadapt=TRUE in \code{make_counts}.
}

\examples{
report <- progress_report(<your_path_to_sports_output_directory>)
report <- progress_report(path=path)

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports. \url{https://github.com/Danis102} for
  updates on the current package.

Other PAC generation: 
\code{\link{import_sports}()},
\code{\link{make_PAC}()},
\code{\link{make_anno}()},
\code{\link{make_counts_dep}()},
\code{\link{make_counts}()},
\code{\link{make_cutadapt}()},
\code{\link{make_pheno}()},
\code{\link{make_trim}()}
}
\concept{PAC generation}
