% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_report.R
\name{progress_report}
\alias{progress_report}
\title{Genereate snRNA Progress Report from Sports}
\usage{
progress_report(path)
}
\arguments{
\item{path}{The Sport output path. The function will search this path for the
the progress_report folder that is automatically generated by Sports
running with the -k option (save all intermediate files).}
}
\value{
data.frame
}
\description{
\code{progress.report} genereates summarized snRNA progress report
from Sports intermediate files.
}
\details{
Given the path to Sports output directory, this function will import all
progress report txt.files of a given Sports project and summarize them into a
single dataframe. The progress report dataframe can then easily be added to
the Pheno object in the PAC file for further analysis.
}
\section{Important}{
 Progress reports will only be generated if using the -k
  option in Sports.
}

\examples{
report <- progress_report(<your_path_to_sports_output_directory>)
report <- progress_report(path=path)

}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
  documentation about running Sports. \url{https://github.com/Danis102} for
  updates on the current package.

Other PAC generation: \code{\link{import_sports}},
  \code{\link{make_anno}}, \code{\link{make_countTable}},
  \code{\link{make_pheno}}
}
\concept{PAC generation}
