% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_S4classes.R
\docType{class}
\name{reanno-class}
\alias{reanno-class}
\alias{reanno}
\title{S4 object for reanno output}
\usage{
reanno(Overview, Full_anno)
}
\value{
Contains two slots:  1. Overview: A table holding a summary of the
mapping. 2. Full_anno: Lists of tables holding the full imports per mismatch
cycle (mis0, mis1 etc) and reference (mi0-ref1, mis0-ref2, mis1-ref1,
mis1-ref2 etc).
}
\description{
Holds the imported information from mapping using the map_reanno function.
All information are held in tibble class (tbl/tbl_df) tables from the tibble
package.
}
