% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{rownames,PAC-method}
\alias{rownames,PAC-method}
\alias{rownames,reanno-method}
\title{sequence names}
\usage{
\S4method{rownames}{PAC}(x)

\S4method{rownames}{reanno}(x)
}
\arguments{
\item{x}{A S4 object of class PAC.}
}
\value{
The sequences (row names) held by the Counts/Anno tables of a
  PAC-object.

Sequences in the reanno-object.
}
\description{
sequence names

sequences in reanno
}
\examples{

load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))

# extra fuctionality with s4 PAC-object:
names(pac)
length(pac)
nrow(pac)
ncol(pac)
rownames(pac)
colnames(pac)
pheno(pac) 
head(anno(pac))
head(counts(pac))
head(norm(pac)$cpm)

}
