% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{show.PAC}
\alias{show.PAC}
\title{show(PAC)}
\usage{
show.PAC(object)
}
\arguments{
\item{object}{A S4 object of class PAC.}
}
\value{
Simple statistical overview of a PAC-object.
}
\description{
Overview (show method) of a S4 PAC
}
\examples{

load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))

# extra fuctionality with s4 PAC-object:
pac_s4 <- as.PAC(pac)
pac_s4
names(pac_s4)
length(pac_s4)
nrow(pac_s4)
ncol(pac_s4)
rownames(pac_s4)
colnames(pac_s4)
pheno(pac_s4) 
head(anno(pac_s4))
head(counts(pac_s4))
head(norm(pac_s4)$cpm)

pac_s4 <- PAC_summary(pac_s4, norm = "cpm", 
                   type = "means", pheno_target=list("stage")) 
pac_s4
head(summary(pac_s4)$cpmMeans_stage)

}
\seealso{
Other PAC methods: 
\code{\link{anno}()},
\code{\link{counts}()},
\code{\link{names,PAC-method}},
\code{\link{norm}()},
\code{\link{pheno}()},
\code{\link{summary}()}
}
\concept{PAC methods}
