% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_reanno.R
\name{simplify_reanno}
\alias{simplify_reanno}
\title{Generates one-hit hierarchial biotype annotation}
\usage{
simplify_reanno(
  input,
  hierarchy,
  mismatches = 2,
  bio_name = "Biotypes",
  PAC_merge = FALSE
)
}
\arguments{
\item{hierarchy}{List with character vectors as biotype search vectors that
will be parsed to grepl() and searched for in the mismatch/biotype
reannotation matrix. Important \emph{hierarchy} is order sensitive. For
example if \emph{hierarchy} is defined as list(rRNA="rRNA", miRNA="miRNA",
piRNA="piRNA") a sequence annotated with 0 mismatches to all three biotypes
will only reported once as rRNA (one-hit) since this biotype was put first
in the list Sequences annotated to neither of the biotypes listed in
\emph{hierarchy} but still is annotated to something will be asigned as
"other". Sequences without an annotation at the allowed mismatch level will
be asigned as "no_anno".}

\item{mismatches}{Integer indicating the number of allowed mismatches. Note
that this value can never be larger than the maximum number of mismatches
used in the reannotation workflow (default=0).}

\item{bio_name}{Character naming the final biotype column (default="Biotype")}

\item{PAC_merge}{Logical whether simplified annotation vector should
automatically be added to Anno object if a PAC list object were given as
input (default=FALSE)}

\item{Input}{Dataframe with biotype/mismatch matrix generated by
\emph{add_reanno}. A PAC-list object can also be provided here where the
Anno dataframe will be used as input.}
}
\value{
Character vector with single best-hit biotypes with fewest mismatches.
}
\description{
\code{simplify_reanno} adds biotypes from reannotation to PAC.
}
\details{
This function is used for dealing with and to explore the effect of
multimapping. Given a mismatch/biotype reannotation matrix (colnames: mis0,
mis1, mis2 etc) created by \emph{add_reanno} and usually added to the Anno
dataframe of a PAC object (colnames: mis0, mis1, mis2 etc) , the function will
generated a one-hit biotype vector according to a user-defined biotype
hierarchy.
}
\examples{

load(file="/home/danis31/OneDrive/Programmering/Programmering/Pipelines/Drosophila/Pipeline_3.1/seqpac/dm_test_PAC.Rdata")

hierarchy <- list( Mt_rRNA= "12S|16S|Mt_rRNA",
                rRNA="5S|5.8S|18S|28S|S45|Ensembl_rRNA|rRNA_Other",
                Mt_tRNA= "tRNA_mt-tRNA",
               tRNA="Ensembl_tRNA|tRNA_nuc-tRNA",
               miRNA="^miRNA|Ensembl_miRNA|Ensembl_pre_miRNA",
               piRNA="piRNA")

PAC_filt <- simplify_reanno(PAC_filt, hierarchy=hierarchy, mismatches=0, bio_name="Biotypes_mis0", PAC_merge=TRUE)
PAC_filt <- simplify_reanno(PAC_filt, hierarchy=hierarchy, mismatches=3, bio_name="Biotypes_mis3", PAC_merge=TRUE)

PAC_check(PAC_filt)


}
\seealso{
\url{https://github.com/junchaoshi/sports1.0} for download and
 documentation about running Sports.
 \url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
 Bowtie \url{https://github.com/Danis102} for updates on the current package.

Other PAC reannotation: 
\code{\link{add_reanno}()},
\code{\link{check_reanno}()},
\code{\link{import_reanno}()},
\code{\link{make_reanno}()},
\code{\link{map_reanno}()}
}
\concept{PAC reannotation}
