% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_reanno.R
\name{simplify_reanno}
\alias{simplify_reanno}
\title{Generates one-hit hierarchial biotype annotation}
\usage{
simplify_reanno(
  input,
  hierarchy,
  mismatches = 2,
  bio_name = "Biotypes",
  merge_pac = FALSE,
  target_columns = NULL
)
}
\arguments{
\item{hierarchy}{List with character vectors. The vectors search terms
constructed as regular expressions that will be parsed to grepl() and
searched for in the mismatch biotype reannotation matrix (colnames:
mis0_bio, mis1_bio, mis2_bio, mis3_bio). Important \emph{hierarchy} is order
sensitive. For example if \emph{hierarchy} is defined as list(rRNA="rRNA",
miRNA="miRNA", piRNA="piRNA") a sequence annotated with 0 mismatches for all
three biotypes will only be reported as rRNA since this biotype was put
first in the list. Sequences annotated to neither of the biotypes listed in
\emph{hierarchy} but still is annotated to something will be asigned as
"other". Sequences without an annotation down to the specified mismatch
level (\code{mismatches} will be asigned as "no_anno".}

\item{mismatches}{Integer indicating the number of allowed mismatches. Note
that this value can never be larger than the maximum number of mismatches
used in the reannotation workflow (default=0).}

\item{bio_name}{Character naming the final biotype column (default="Biotype")}

\item{merge_pac}{Logical whether the simplified annotation column should
automatically be added to the Anno object if a PAC list object were given as
input (default=FALSE). Alternatively, if a PAC object is provided in
merge_pac, the simplified column will be added to the Anno table for this
PAC object.}

\item{Input}{Data.frame with biotype/mismatch matrix generated by
\code{add_reanno}. A PAC-list object can also be provided here where the
Anno data.frame will be used as input. Important, PAC objects need biotype
columns generated by \code{\link{add_reanno}}.}
}
\value{
Character vector with single best-hit biotypes with fewest mismatches.
}
\description{
\code{simplify_reanno} adds hierarchical classification to PAC.
}
\details{
This function deals with and can be used to explores the effect of
multimapping. Given a biotype reannotation matrix (colnames = mis0_bio,
mis1_bio, mis2_bio etc) created by \code{\link{add_reanno}} and usually added
to the Anno data.frame of a PAC object, the function will generated a one-hit
biotype vector according to a user-defined biotype hierarchy.
}
\examples{

# See, vignette(package="seqpac")
# 
##--------------------------------------------------------------------------#
## Hierarchical classification with simplify_reanno
##--------------------------------------------------------------------------#
# table(pac$Anno$mis0_bio)
# 
# hierarchy <- list(rRNA="Ensembl_rRNA",
#                    Mt_tRNA="tRNA:MT",
#                    tRNA="Ensembl_tRNA|tRNA__tRNA",
#                   miRNA ="^miRNA|Ensembl_miRNA|Ensembl_pre_miRNA",
#                    snoRNA="Ensembl_snoRNA",
#                    lncRNA="Ensembl_lncRNA"   
#                   )
# 
# ## No mistmach allowed
# pac <- simplify_reanno(input=pac, hierarchy=hierarchy, mismatches=0, 
#                       bio_name="Biotypes_mis0", merge_pac=TRUE)
# 
# ## Up to 3 mismatches allowed
# pac <- simplify_reanno(input=pac, hierarchy=hierarchy, mismatches=3, 
#                        bio_name="Biotypes_mis3", merge_pac=TRUE)
# 
# # Example of original reference name annotations 
# head(reanno_biotype$Full_anno$mis0$Ensembl)
# 
# # Simplified hierarchical classification
# head(pac$Anno)
# table(pac$Anno$Biotypes_mis0)
# table(pac$Anno$Biotypes_mis3)
#
##--------------------------------------------------------------------------#
## Simplify and make hierarchy of reannotation 
##--------------------------------------------------------------------------#
# library(seqpac)
# load(system.file("extdata", "drosophila_sRNA_pac.Rdata", 
#                   package = "seqpac", mustWork = TRUE))
# 
# # Build hierarchy with regular expressions:
# hierarchy <- list(rRNA="Ensembl_rRNA",
#                   Mt_tRNA="tRNA_mt:tRNA",
#                   tRNA="Ensembl_tRNA|tRNA_\\^tRNA",
#                   miRNA ="^miRNA|Ensembl_miRNA|Ensembl_pre_miRNA"
#                  )
#
# pac_master <- simplify_reanno(pac_master, hierarchy=hierarchy, 
#                               mismatches=0, bio_name="Biotypes_mis0", 
#                               merge_pac=TRUE)
#
# PAC_check(pac_master)


}
\seealso{
\url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
 Bowtie \url{https://github.com/Danis102} for updates on the current package.

Other PAC reannotation: 
\code{\link{add_reanno}()},
\code{\link{check_reanno}()},
\code{\link{import_reanno}()},
\code{\link{make_reanno}()},
\code{\link{map_reanno}()}
}
\concept{PAC reannotation}
