% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_reanno.R
\name{simplify_reanno}
\alias{simplify_reanno}
\title{Generates one-hit hierarchial biotype annotation}
\usage{
simplify_reanno(
  input,
  hierarchy,
  mismatches = 2,
  bio_name = "Biotypes",
  merge_pac = FALSE,
  target_columns = NULL
)
}
\arguments{
\item{hierarchy}{List with character vectors. The vectors search terms
constructed as regular expressions that will be parsed to grepl() and
searched for in the mismatch/biotype reannotation matrix. Important
\emph{hierarchy} is order sensitive. For example if \emph{hierarchy} is
defined as list(rRNA="rRNA", miRNA="miRNA", piRNA="piRNA") a sequence
annotated with 0 mismatches for all three biotypes will only be reported as
rRNA since this biotype was put first in the list. Sequences annotated to
neither of the biotypes listed in \emph{hierarchy} but still is annotated to
something will be asigned as "other". Sequences without an annotation down
to the specified mismatch level (\code{mismatches} will be asigned as
"no_anno".}

\item{mismatches}{Integer indicating the number of allowed mismatches. Note
that this value can never be larger than the maximum number of mismatches
used in the reannotation workflow (default=0).}

\item{bio_name}{Character naming the final biotype column (default="Biotype")}

\item{merge_pac}{Logical whether the simplified annotation column should
automatically be added to the Anno object if a PAC list object were given as
input (default=FALSE). Alternatively, if a PAC object is provided in
merge_pac, the simplified column will be added to the Anno table for this
PAC object.}

\item{Input}{Dataframe with biotype/mismatch matrix generated by
\code{add_reanno}. A PAC-list object can also be provided here where the
Anno dataframe will be used as input.}
}
\value{
Character vector with single best-hit biotypes with fewest mismatches.
}
\description{
\code{simplify_reanno} adds biotypes from reannotation to PAC.
}
\details{
This function deals with and can be used to explores the effect of
multimapping. Given a biotype reannotation matrix (colnames: mis0_bio,
mis1_bio, mis2_bio etc) created by \code{add_reanno} and usually added to the
Anno dataframe of a PAC object, the function will generated a one-hit biotype
vector according to a user-defined biotype hierarchy.
}
\examples{


#-------------------------------------------------------------------------------------------------#
# Simplify and make hierarchy of reannotation 
#-------------------------------------------------------------------------------------------------#
library(seqpac)
load(system.file("extdata", "drosophila_sRNA_pac.Rdata", package = "seqpac", mustWork = TRUE))

# Build hierarchy with regular expressions:
hierarchy <- list(rRNA="Ensembl_rRNA",
                  Mt_tRNA="tRNA_mt:tRNA",
                  tRNA="Ensembl_tRNA|tRNA_^tRNA",
                  miRNA ="^miRNA|Ensembl_miRNA|Ensembl_pre_miRNA"
                 )

PAC_filt <- simplify_reanno(pac_master, hierarchy=hierarchy, mismatches=0, bio_name="Biotypes_mis0", PAC_merge=TRUE)
PAC_filt <- simplify_reanno(PAC_filt, hierarchy=hierarchy, mismatches=3, bio_name="Biotypes_mis3", PAC_merge=TRUE)

PAC_check(PAC_filt)


}
\seealso{
\url{http://bowtie-bio.sourceforge.net/index.shtml} for information about
 Bowtie \url{https://github.com/Danis102} for updates on the current package.

Other PAC reannotation: 
\code{\link{add_reanno}()},
\code{\link{check_reanno}()},
\code{\link{import_reanno}()},
\code{\link{make_reanno}()},
\code{\link{map_reanno}()}
}
\concept{PAC reannotation}
