% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{summary}
\alias{summary}
\alias{summary<-}
\alias{summary,PAC-method}
\title{list of summary tables}
\usage{
summary(object)

summary(object) <- value

\S4method{summary}{PAC}(object)
}
\description{
list of summary tables
}
