% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpac_generics_methods.R
\name{summary}
\alias{summary}
\alias{summary<-}
\alias{summary,PAC-method}
\title{list of summary tables}
\usage{
summary(object)

summary() <- value

\S4method{summary}{PAC}(object)
}
\value{
The list of summary tables from a PAC-object
}
\description{
list of summary tables
}
\examples{

load(system.file("extdata", "drosophila_sRNA_pac_filt_anno.Rdata", 
                  package = "seqpac", mustWork = TRUE))

# extra fuctionality with s4 PAC-object:
pac_s4 <- as.PAC(pac)
pac_s4
names(pac_s4)
length(pac_s4)
nrow(pac_s4)
ncol(pac_s4)
rownames(pac_s4)
colnames(pac_s4)
pheno(pac_s4) 
head(anno(pac_s4))
head(counts(pac_s4))
head(norm(pac_s4)$cpm)

pac_s4 <- PAC_summary(pac_s4, norm = "cpm", 
                   type = "means", pheno_target=list("stage")) 
pac_s4                   
head(summary(pac_s4)$cpmMeans_stage)

}
