% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_class.R
\name{tRNA_class}
\alias{tRNA_class}
\title{Classify tsRNA/tRFs from  PAC object}
\usage{
tRNA_class(PAC, map, terminal)
}
\arguments{
\item{PAC}{PAC-list object.}

\item{map}{Map object with loop annotations generated by
\code{\link[seqpac]{PAC_mapper}} followed by
\code{\link[seqpac]{map_rangetype}} functions. Note, only full length
reference tRNAs annotated with 3 loops in the ss file (see:
\code{\link[seqpac]{map_rangetype}}) are used in the classifications.
Explanation: sometimes tRNAscan-SE scans results in ss files with
incomplete secondary structures with either more or fewer loops than the D,
T and anticodon loops found in canonical tRNAs. Such tRNAs will be reomved
from the analysis.}

\item{terminal}{Integer setting the terminal tolerance. Example, when terminal
is set to 5 (default) sequenced reads starting within the 5 first
nucleotides of the full-length tRNA reference will be classifed as 5'. If
it instead ends within the 5 last nucleotides of the full-length tRNA
reference it will be classified as 3'. If a read neither starts nor ends
within the terminal threshold it will be classified as i' (internal). Note,
for tsRNA/tRF classification we recommend a threshold of terminal=2 (not
default). Advisably, however, \code{\link[seqpac]{PAC_mapper}} runs with
N_up = "NNN" and N_down = "NNN" when mapping tsRNA/tRFs. Thus, the default
threshold of 5 is used when full-length tRNAs has been extended with NNN
up- and downstream to compensate for possible fragments from mature tRNAs
(3+2). Always double check the exact fragment using
\code{\link[seqpac]{PAC_covplot}}.}
}
\value{
Merged PAC object with an extended Anno table containing the
  tRF/tsRNA classifications.
}
\description{
Classifying tRFs/tsRNA into 5'-tRF/halves, 3'-tRFs/halves, i'-tRFs/halves, as
well as acceptor and decoder isotypes.
}
\details{
Given a map object with range types generated by
\code{\link[seqpac]{PAC_mapper}} followed by
\code{\link[seqpac]{map_rangetype}} functions, sequences in a PAC object are
classified according the terminals (5'/3'/i'), anticodon loop (half/tRF), and
isotype (decoder/accepter) of the full length tRNA.
}
\seealso{
\url{https://github.com/Danis102} for updates on the current
  package.

Other PAC analysis: 
\code{\link{PAC_covplot}()},
\code{\link{PAC_deseq}()},
\code{\link{PAC_filter}()},
\code{\link{PAC_filtsep}()},
\code{\link{PAC_gtf}()},
\code{\link{PAC_jitter}()},
\code{\link{PAC_mapper}()},
\code{\link{PAC_mismap}()},
\code{\link{PAC_nbias}()},
\code{\link{PAC_norm}()},
\code{\link{PAC_pca}()},
\code{\link{PAC_pie}()},
\code{\link{PAC_pirna}()},
\code{\link{PAC_rpm}()},
\code{\link{PAC_saturation}()},
\code{\link{PAC_sizedist}()},
\code{\link{PAC_stackbar}()},
\code{\link{PAC_summary}()},
\code{\link{PAC_trna}()},
\code{\link{map_rangetype}()}
}
\concept{PAC analysis}
